% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{align_transform}
\alias{align_transform}
\title{align_transform is not supported. Co-embed your data sets if
you need batch correction.}
\usage{
align_transform(cds, reduction_method = c("Aligned"))
}
\description{
align_transform is not supported. Co-embed your data sets if
you need batch correction.
}
