% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_cells.R
\name{branch_nodes}
\alias{branch_nodes}
\title{Return the names of principal graph nodes that are branches (excluding roots)}
\usage{
branch_nodes(cds, reduction_method = "UMAP")
}
\description{
Return the names of principal graph nodes that are branches (excluding roots)
}
