% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l1graph.R
\name{calc_principal_graph}
\alias{calc_principal_graph}
\title{Function to automatically learn the structure of data by either using L1-graph or the spanning-tree formulization}
\usage{
calc_principal_graph(X, C0, G, maxiter = 10, eps = 1e-05,
  gstruct = c("l1-graph", "span-tree"), L1.lambda = 1,
  L1.gamma = 0.5, L1.sigma = 0.01, nn = 5, L1.timeout = 1800,
  verbose = T)
}
\arguments{
\item{X}{the input data DxN}

\item{C0}{the initialization of centroids}

\item{G}{graph matrix with side information where cannot-link pair is 0}

\item{maxiter}{maximum number of iteraction}

\item{eps}{relative objective difference}

\item{gstruct}{graph structure to learn, either L1-graph or the span-tree}

\item{L1.lambda}{regularization parameter for inverse graph embedding}

\item{L1.gamma}{regularization parameter for k-means (the prefix of 'param' is used to avoid name collision with gamma)}

\item{L1.sigma}{bandwidth parameter}

\item{nn}{number of nearest neighbors}

\item{L1.timeout}{a positive integer value specifying the number of seconds after which a timeout will occur. If zero, then no timeout will occur. (This is a parameter passed to lp.control function)}

\item{verbose}{emit results from iteraction}
}
\value{
a list of X, C, W, P, objs
X is the input data
C is the centers for principal graph
W is the pricipal graph matrix
P is the cluster assignment matrix
objs is the objective value for the function
}
\description{
Function to automatically learn the structure of data by either using L1-graph or the spanning-tree formulization
}
