% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_data_set.R
\docType{class}
\name{cell_data_set}
\alias{cell_data_set}
\alias{cell_data_set-class}
\title{The cell_data_set class}
\description{
The main class used by Monocle3 to hold single cell expression data.
cell_data_set extends the basic Bioconductor SingleCellExperiment class.
}
\details{
This class is initialized from a matrix of expression values Methods that
operate on cell_data_set objects constitute the basic Monocle workflow.
}
