% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_cells.R
\name{choose_graph_segment}
\alias{choose_graph_segment}
\title{Choose cells interactively to make subset cds}
\usage{
choose_graph_segment(cds, reduction_method = "UMAP",
  return_list = FALSE)
}
\arguments{
\item{cds}{CDS object}

\item{return_list}{Logical, return a list of cells.}
}
\value{
A subset CDS object. If return_list = FALSE, a list of cell names.
}
\description{
Choose cells interactively to make subset cds
}
