% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clear_cds_slots}
\alias{clear_cds_slots}
\title{Clear CDS slots}
\usage{
clear_cds_slots(cds)
}
\arguments{
\item{cds}{cell_data_set to be cleared}
}
\value{
A cell_data_set with only expression, rowData and colData present.
}
\description{
Function to clear all CDS slots besides colData, rowData and expression data.
}
