% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_cells.R
\name{cluster_cells}
\alias{cluster_cells}
\title{Cluster cells into a specified number of groups based on .}
\usage{
cluster_cells(cds, reduced_dimension = c("tSNE", "UMAP", "PCA"),
  use_pca = FALSE, k = 20, louvain_iter = 1, weight = FALSE,
  res = NULL, method = c("louvain"), random_seed = 0L, verbose = F,
  cores = 1, ...)
}
\arguments{
\item{cds}{the cell_data_set upon which to perform this operation}

\item{use_pca}{Whether or not to cluster cells based on top PCA component. Default to be FALSE.}

\item{k}{number of kNN used in creating the k nearest neighbor graph for Louvain clustering. The number of kNN is related to the resolution of the clustering result, bigger number of kNN gives low resolution and vice versa. Default to be 20}

\item{louvain_iter}{Integer number of iterations used for Louvain clustering. The clustering result gives the largest modularity score will be used as the final clustering result.  Default to be 1. Note that if louvain_iter is large than 1, the \code{seed} argument will be ignored.}

\item{weight}{A logic argument to determine whether or not we will use Jaccard coefficent for two nearest neighbors (based on the overlapping of their kNN) as the weight used for Louvain clustering. Default to be FALSE.}

\item{res}{Resolution parameter for the louvain clustering. Values between 0 and 1e-2 are good, bigger values give you more clusters. Default is set to be \code{seq(0, 1e-4, length.out = 5)}.}

\item{method}{method for clustering cells. Three methods are available, including densityPeak, louvian and DDRTree. By default, we use density peak clustering algorithm for clustering. For big datasets (like data with 50 k cells or so), we recommend using the louvain clustering algorithm.}

\item{random_seed}{the seed used by the random number generator in louvain-igraph package. This argument will be ignored if louvain_iter is larger than 1.}

\item{verbose}{Verbose A logic flag to determine whether or not we should print the running details.}

\item{cores}{number of cores computer should use to execute function}

\item{...}{Additional arguments passed to \code{\link{densityClust}()}}
}
\value{
an updated cell_data_set object, in which phenoData contains values for Cluster for each cell
}
\description{
Unsupervised clustering of cells is a common step in many single-cell expression
workflows. In an experiment containing a mixture of cell types, each cluster might
correspond to a different cell type. This method takes a cell_data_set as input
along with a requested number of clusters, clusters them with an unsupervised
algorithm (by default, density peak clustering), and then returns the cell_data_set with the
cluster assignments stored in the colData table. When number of clusters is set
to NULL (num_clusters = NULL), the decision plot as introduced in the reference
will be plotted and the users are required to check the decision plot to select
the rho and delta to determine the number of clusters to cluster. When the dataset
is big, for example > 50 k, we recommend the user to use the Louvain clustering
algorithm which is inspired from phenograph paper. Note Louvain doesn't support the
num_cluster argument but the k (number of k-nearest neighbors) is relevant to the final
clustering number. The deafult implementation of Louvain clustering (when res is set to be NULL)
is based on the Rphenograph package but updated based on our requirement (for example,
changed the jaccard_coeff function as well as adding louvain_iter argument, etc.). We also
support setting resolutioin parameter when performing louvain clustering using the louvain package
from python. With different res values, the users can obtain different granularity of the data.
}
\references{
Rodriguez, A., & Laio, A. (2014). Clustering by fast search and find of density peaks. Science, 344(6191), 1492-1496. doi:10.1126/science.1242072

Vincent D. Blondel, Jean-Loup Guillaume, Renaud Lambiotte, Etienne Lefebvre: Fast unfolding of communities in large networks. J. Stat. Mech. (2008) P10008

Jacob H. Levine and et.al. Data-Driven Phenotypic Dissection of AML Reveals Progenitor-like Cells that Correlate with Prognosis. Cell, 2015.
}
