% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_genes.R
\name{cluster_genes}
\alias{cluster_genes}
\title{Cluster genes into a specified number of groups based on .}
\usage{
cluster_genes(cds, reduction_method = c("UMAP", "tSNE", "PCA"),
  preprocess_method = c("PCA", "LSI"), max_components = 2,
  umap.metric = "cosine", umap.min_dist = 0.1,
  umap.n_neighbors = 15L, umap.fast_sgd = TRUE,
  umap.nn_method = "annoy", k = 20, louvain_iter = 1,
  louvain_qval = 0.05, weight = FALSE, resolution = NULL,
  random_seed = 0L, cores = 1, verbose = F, ...)
}
\arguments{
\item{cds}{the cell_data_set upon which to perform this operation}

\item{k}{number of kNN used in creating the k nearest neighbor graph for Louvain clustering. The number of kNN is related to the resolution of the clustering result, bigger number of kNN gives low resolution and vice versa. Default to be 20}

\item{louvain_iter}{Integer number of iterations used for Louvain clustering. The clustering result gives the largest modularity score will be used as the final clustering result.  Default to be 1. Note that if louvain_iter is large than 1, the \code{seed} argument will be ignored.}

\item{weight}{A logic argument to determine whether or not we will use Jaccard coefficent for two nearest neighbors (based on the overlapping of their kNN) as the weight used for Louvain clustering. Default to be FALSE.}

\item{random_seed}{the seed used by the random number generator in louvain-igraph package. This argument will be ignored if louvain_iter is larger than 1.}

\item{cores}{number of cores computer should use to execute function}

\item{verbose}{Verbose A logic flag to determine whether or not we should print the running details.}

\item{res}{Resolution parameter for the louvain clustering. Values between 0 and 1e-2 are good, bigger values give you more clusters. Default is set to be \code{seq(0, 1e-4, length.out = 5)}.}
}
\value{
an updated cell_data_set object, in which phenoData contains values for Cluster for each cell
}
\description{
Cluster genes into a specified number of groups based on .
}
\references{
Rodriguez, A., & Laio, A. (2014). Clustering by fast search and find of density peaks. Science, 344(6191), 1492-1496. doi:10.1126/science.1242072

Vincent D. Blondel, Jean-Loup Guillaume, Renaud Lambiotte, Etienne Lefebvre: Fast unfolding of communities in large networks. J. Stat. Mech. (2008) P10008

Jacob H. Levine and et.al. Data-Driven Phenotypic Dissection of AML Reveals Progenitor-like Cells that Correlate with Prognosis. Cell, 2015.
}
