% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{clusters,cell_data_set-method}
\alias{clusters,cell_data_set-method}
\title{Method to extract clusters from CDS object}
\usage{
\S4method{clusters}{cell_data_set}(x, reduction_method = "UMAP")
}
\arguments{
\item{x}{A cell_data_set object.}

\item{reduction_method}{Reduced dimension to extract clusters for.}
}
\value{
Clusters.
}
\description{
Method to extract clusters from CDS object
}
