% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{clusters}
\alias{clusters}
\title{Generic to extract clusters from CDS object}
\usage{
clusters(x, reduction_method = "UMAP")
}
\arguments{
\item{x}{A cell_data_set object.}

\item{reduction_method}{Reduced dimension to extract clusters for.}
}
\description{
Generic to extract clusters from CDS object
}
\examples{
  \donttest{
    cds <- load_worm_embryo()
    cds <- preprocess_cds(cds)
    cds <- reduce_dimension(cds)
    cds <- cluster_cells(cds)
    clusters_factors <- clusters(cds, "UMAP")
  }

}
