% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{convert_counts_matrix}
\alias{convert_counts_matrix}
\title{Convert the counts matrix class in the given CDS.}
\usage{
convert_counts_matrix(cds, matrix_control = list())
}
\arguments{
\item{cds}{cell_data_set The cell_data_set that has the
counts matrix to be converted.}

\item{matrix_control}{list A list of matrix control
values used to convert the counts matrix. If the
counts matrix in the cds is the same as the desired
counts matrix, it is not altered. matrix_control is
required.}
}
\value{
cell_data_set The cell_data_set with the converted
counts matrix.
}
\description{
Converts the counts matrix that is in the
given CDS to the matrix_class specified in the
matrix_control list.
}
\examples{
   cds <- load_a549()
   str(counts(cds))
   cds <- convert_counts_matrix(cds, matrix_control=list(matrix_class='BPCells'))
   str(counts(cds))

}
