% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{counts_row_order}
\alias{counts_row_order}
\title{Generic to access cds row order BPCells counts matrix.}
\usage{
counts_row_order(x)
}
\arguments{
\item{x}{A cell_data_set object.}
}
\value{
BPCells row order counts matrix.
}
\description{
Generic to access cds row order BPCells counts matrix.
}
\examples{
 \donttest{
   if (requireNamespace("BPCells", quietly = TRUE)) {
     cds <- load_a549(matrix_control=list(matrix_class='BPCells'))
     mat_row_order <- counts_row_order(cds)
   }
 }         
         
}
