% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_dispersions.R
\name{estimate_dispersions_cds}
\alias{estimate_dispersions_cds}
\title{Helper function to estimate dispersions}
\usage{
estimate_dispersions_cds(cds, modelFormulaStr, relative_expr,
  min_cells_detected, removeOutliers, verbose = FALSE)
}
\arguments{
\item{cds}{a cell_data_set that contains all cells user wants evaluated}

\item{modelFormulaStr}{a formula string specifying the model to fit for the genes.}

\item{relative_expr}{Whether to transform expression into relative values}

\item{min_cells_detected}{Only include genes detected above lower_detection_limit in at least this many cells in the dispersion calculation}

\item{removeOutliers}{a boolean it determines whether or not outliers from the data should be removed}

\item{verbose}{Whether to show detailed running information.}
}
\description{
Helper function to estimate dispersions
}
