% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_models.R
\name{evaluate_fits}
\alias{evaluate_fits}
\title{Evaluate the fits of model objects.}
\usage{
evaluate_fits(model_tbl)
}
\arguments{
\item{model_tbl}{A tibble of model objects, generally output of
\code{\link{fit_models}}.}
}
\description{
Evaluate the fits of model objects.
}
