% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{exprs,cell_data_set-method}
\alias{exprs,cell_data_set-method}
\title{Method to access cds count matrix}
\usage{
\S4method{exprs}{cell_data_set}(x)
}
\arguments{
\item{x}{A cell_data_set object.}
}
\description{
Method to access cds count matrix
}
