% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{exprs}
\alias{exprs}
\title{Generic to access cds count matrix}
\usage{
exprs(x)
}
\arguments{
\item{x}{A cell_data_set object.}
}
\value{
Count matrix.
}
\description{
Generic to access cds count matrix
}
\examples{
 \donttest{
   cds <- load_a549()
   exprs(cds)
 }

}
