% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{fData,cell_data_set-method}
\alias{fData,cell_data_set-method}
\title{Generic to access cds rowData table}
\usage{
\S4method{fData}{cell_data_set}(x)
}
\arguments{
\item{x}{A cell_data_set object.}
}
\description{
Generic to access cds rowData table
}
