% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{fData<-,cell_data_set-method}
\alias{fData<-,cell_data_set-method}
\title{Method to set cds rowData table}
\usage{
\S4method{fData}{cell_data_set}(x) <- value
}
\arguments{
\item{x}{A cell_data_set object.}

\item{value}{A data frame to set to colData table.}
}
\description{
Method to set cds rowData table
}
\examples{
  \donttest {
    cds <- load_a549()
    fData(cds)[['row_index']] <- seq(nrow(fData(cds)))
  }

}
