% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{fData<-}
\alias{fData<-}
\title{Generic to set cds rowData table}
\usage{
fData(x) <- value
}
\arguments{
\item{x}{A cell_data_set object.}

\item{value}{A data frame to set to colData table.}
}
\description{
Generic to set cds rowData table
}
