% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_graph.R
\name{find_nearest_vertex}
\alias{find_nearest_vertex}
\title{Finds the nearest principal graph node}
\usage{
find_nearest_vertex(data_matrix, target_points, block_size = 50000,
  process_targets_in_blocks = FALSE)
}
\arguments{
\item{data_matrix}{the input matrix}

\item{target_points}{the target points}

\item{block_size}{the number of input matrix rows to process per block}

\item{process_targets_in_blocks}{whether to process the targets points in
blocks instead}
}
\description{
Finds the nearest principal graph node
}
\keyword{internal}
