% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_graph.R
\name{findNearestVertex}
\alias{findNearestVertex}
\title{Finds the nearest principal graph node}
\usage{
findNearestVertex(data_matrix, target_points, block_size = 50000,
  process_targets_in_blocks = FALSE)

findNearestVertex(data_matrix, target_points, block_size = 50000,
  process_targets_in_blocks = FALSE)
}
\arguments{
\item{data_matrix}{the input matrix}

\item{target_points}{the target points}

\item{block_size}{the number of input matrix rows to process per bloclk}

\item{process_targets_in_blocks}{whether to process the targets points in blocks instead}

\item{data_matrix}{the input matrix}

\item{target_points}{the target points}

\item{block_size}{the number of input matrix rows to process per bloclk}

\item{process_targets_in_blocks}{whether to process the targets points in blocks instead}
}
\description{
Finds the nearest principal graph node

Finds the nearest principal graph node
}
\keyword{internal}
