% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_models.R
\name{fit_models}
\alias{fit_models}
\title{Fits a model for each gene in a cell_data_set object.}
\usage{
fit_models(
  cds,
  model_formula_str,
  expression_family = "quasipoisson",
  reduction_method = "UMAP",
  cores = 1,
  clean_model = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{cds}{The cell_data_set upon which to perform this operation.}

\item{model_formula_str}{A formula string specifying the model to fit for
the genes.}

\item{expression_family}{Specifies the family function used for expression
responses. Can be one of "quasipoisson", "negbinomial", "poisson",
"binomial", "gaussian", "zipoisson", or "zinegbinomial". Default is
"quasipoisson".}

\item{reduction_method}{Which method to use with clusters() and
partitions(). Default is "UMAP".}

\item{cores}{The number of processor cores to use during fitting.}

\item{clean_model}{Logical indicating whether to clean the model. Default is
TRUE.}

\item{verbose}{Logical indicating whether to emit progress messages.}

\item{...}{Additional arguments passed to model fitting functions.}
}
\value{
a tibble where the rows are genes and columns are
\itemize{
\item id from \code{rowData(cds)$id}.
\item gene_short_names from \code{rowData(cds)$gene_short_names}.
\item num_cells_expressed from \code{rowData(cds)$num_cells_expressed}.
\item gene_id from \code{row.names(rowData(cds))}.
\item model  GLM model returned by speedglm.
\item model_summary returned by \code{summary(model)}.
\item status of model fitting: OK when model converged, otherwise FAIL.
}
}
\description{
This function fits a generalized linear model for each gene in a
cell_data_set. Formulae can be provided to account for additional covariates
(e.g. day collected, genotype of cells, media conditions, etc).
}
