% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_transfer.R
\name{fix_missing_cell_labels}
\alias{fix_missing_cell_labels}
\title{export}
\usage{
fix_missing_cell_labels(
  cds,
  reduction_method = c("UMAP", "PCA"),
  out_notna_model_dir = NULL,
  cell_label_type = NULL,
  k = 10,
  nn_control = nn_control,
  top_threshold = 0.5,
  top_over_second_threshold = 1.5,
  verbose = FALSE,
  ...
)
}
\description{
export
}
