% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_graph.R
\name{generate_centers}
\alias{generate_centers}
\title{Function to reproduce the behavior of eye function in matlab}
\usage{
generate_centers(X, W, P, param.gamma)
}
\arguments{
\item{X}{input data}

\item{W}{the pricipal graph matrix}

\item{P}{the cluster assignment matrix}

\item{param.gamma}{regularization parameter for k-means (the prefix of
'param' is used to avoid name collision with gamma)}
}
\value{
A matrix C for the centers for principal graph
}
\description{
Function to reproduce the behavior of eye function in matlab
}
