% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_citations}
\alias{get_citations}
\title{Access citations for methods used during analysis.}
\usage{
get_citations(cds)
}
\arguments{
\item{cds}{The cds object to access citations from.}
}
\value{
A data frame with the methods used and the papers to be cited.
}
\description{
Access citations for methods used during analysis.
}
\examples{
{
\dontrun{
get_citations(cds)
}
}
}
