% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{get_clusters,cell_data_set-method}
\alias{get_clusters,cell_data_set-method}
\title{Method to extract clusters from CDS object}
\usage{
\S4method{get_clusters}{cell_data_set}(x, reduction_method = "UMAP")
}
\arguments{
\item{cell_data_set}{CDS object}
}
\description{
Method to extract clusters from CDS object
}
