% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_clusters}
\alias{get_clusters}
\title{Generic to extract clusters from CDS object}
\usage{
get_clusters(x, reduction_method = "UMAP")
}
\arguments{
\item{x}{CDS object}

\item{reduction_method}{Reduced dimension to extract clusters for.}
}
\description{
Generic to extract clusters from CDS object
}
