% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_cellranger_data.R
\name{get_genome_in_matrix_path}
\alias{get_genome_in_matrix_path}
\title{Get a genome from Cell Ranger output}
\usage{
get_genome_in_matrix_path(matrix_path, genome = NULL)
}
\arguments{
\item{matrix_path}{Path to a matrices directory produced by the Cell Ranger pipeline}

\item{genome}{Genome to specifically check for, otherwise will check for whatever genome(s) exist there}
}
\value{
A string representing the genome found
}
\description{
Get a genome from Cell Ranger output
}
