% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_partitions}
\alias{get_partitions}
\title{Generic to extract partitions from CDS object}
\usage{
get_partitions(x, reduction_method = "UMAP")
}
\arguments{
\item{x}{CDS object}

\item{reduction_method}{Reduced dimension to extract clusters for.}
}
\description{
Generic to extract partitions from CDS object
}
