% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identity.R
\name{identity_table}
\alias{identity_table}
\title{Report the matrix and model identity information.}
\usage{
identity_table(cds)
}
\arguments{
\item{cds}{the cell_data_set to use.}
}
\value{
Write identity information to stdout.
}
\description{
Write the cell_data_set matrix and model identity
information to stdout.
}
