% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_cells.R
\name{leaf_nodes}
\alias{leaf_nodes}
\title{Return the names of principal graph nodes that are leaves (excluding roots)}
\usage{
leaf_nodes(cds, reduction_method = "UMAP")
}
\description{
Return the names of principal graph nodes that are leaves (excluding roots)
}
