% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load_align_cds_model}
\alias{load_align_cds_model}
\title{Load align_cds model.}
\usage{
load_align_cds_model(cds, file_name_root = NULL)
}
\arguments{
\item{cds}{cell_data_set into which the model is to be loaded.}

\item{file_name_root.}{A string with the root of names given to the two
files to which the model was saved. The files are named
<file_name_root>.aln_rds
and <file_name_root>.aln_nn_index<n>.}
}
\value{
a cell_data_set
}
\description{
Load into an existing cell_data_set an align_cds model consisting
of dimensionality reduction models and nearest neighbor index
objects, which are used to transform new data sets.
}
