% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_lung}
\alias{load_lung}
\title{Build a cell_data_set from the data stored in inst/extdata directory.}
\usage{
load_lung()
}
\description{
Build a cell_data_set from the data stored in inst/extdata directory.
}
