% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retired.R
\name{load_monocle_objects_retired}
\alias{load_monocle_objects_retired}
\title{Load Monocle3 cell_data_set objects.}
\usage{
load_monocle_objects_retired(directory_path)
}
\arguments{
\item{directory_path}{A string giving the directory from
which to read the previously saved cell_data_set files.}
}
\value{
A cell_data_set.
}
\description{
Load a full Monocle3 cell_data_set from the directory
made by the previously run save_monocle_objects function.
}
