% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load_mtx_data}
\alias{load_mtx_data}
\title{Load data from matrix market format}
\usage{
load_mtx_data(
  mat_path,
  gene_anno_path,
  cell_anno_path,
  umi_cutoff = 100,
  matrix_control = list()
)
}
\arguments{
\item{mat_path}{Path to the .mtx matrix market file.}

\item{gene_anno_path}{Path to gene annotation file.}

\item{cell_anno_path}{Path to cell annotation file.}

\item{umi_cutoff}{UMI per cell cutoff, default is 100.}

\item{matrix_control}{A list used to control how the counts matrix is stored
in the CDS. By default, Monocle3 stores the counts matrix in memory as a
sparse matrix. Setting 'matrix_control=list(matrix_class="BPCells")',
stores the matrix on disk as a sparse matrix.}
}
\value{
cds object
}
\description{
Load data from matrix market format
}
\examples{
  \donttest{
    pmat<-system.file("extdata", "matrix.mtx.gz", package = "monocle3")
    prow<-system.file("extdata", "features_c3h0.txt", package = "monocle3")
    pcol<-system.file("extdata", "barcodes_c2h0.txt", package = "monocle3")
    cds <- load_mtx_data( pmat, prow, pcol)
  }

}
