% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_preprocess_classifier}
\alias{load_preprocess_classifier}
\title{Load preprocess classifier.}
\usage{
load_preprocess_classifier(cds, file_root_name = NULL)
}
\arguments{
\item{cds}{cell_data_set into which the classifier is to be loaded.}

\item{file_root_name.}{A string giving the root of names given to the two
files to which the classifier was saved. The files are named <file_root_name>.rds
and <file_root_name>.annoy_index.}
}
\value{
None
}
\description{
Load into an existing cell_data_set a preprocess classifier consisting
of dimensionality reduction model and Annoy nearest neighbor data
structures, which are used to transform and classify new data sets.
}
