% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_preprocess_model}
\alias{load_preprocess_model}
\title{Load preprocess model.}
\usage{
load_preprocess_model(cds, file_root_name = NULL)
}
\arguments{
\item{cds}{cell_data_set into which the model is to be loaded.}

\item{file_root_name.}{A string giving the root of names given to the two
files to which the model was saved. The files are named <file_root_name>.rds
and <file_root_name>.ppc_nn_index.}
}
\value{
None
}
\description{
Load into an existing cell_data_set a preprocess model consisting
of dimensionality reduction model and nearest neighbor nearest neighbor index
objects, which are used to transform new data sets.
}
