% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retired.R
\name{load_preprocess_model_retired}
\alias{load_preprocess_model_retired}
\title{Load preprocess model.}
\usage{
load_preprocess_model_retired(cds, file_name_root = NULL)
}
\arguments{
\item{cds}{cell_data_set into which the model is to be loaded.}

\item{file_name_root.}{A string with the root of names given to the two
files to which the model was saved. The files are named
<file_name_root>.ppc_rds and <file_name_root>.ppc_nn_index.}
}
\value{
a cell_data_set
}
\description{
Load into an existing cell_data_set a preprocess model consisting
of a dimensionality reduction model and nearest neighbor nearest neighbor index
objects, which are used to transform new data sets.
}
