% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_reduce_dimension_model}
\alias{load_reduce_dimension_model}
\title{Load reduce_dimension model.}
\usage{
load_reduce_dimension_model(cds, file_root_name = NULL)
}
\arguments{
\item{cds}{cell_data_set into which the model is to be loaded.}

\item{file_root_name.}{A string giving the root of names given to the two
files to which the model was saved. The files are named <file_root_name>.rds,
<file_root_name>._umap_nn_index, and <file_root_name>._rdd_nn_index.}
}
\value{
None
}
\description{
Load into an existing cell_data_set a reduce_dimension model consisting
of dimensionality reduction models and nearest neighbor index objects,
which are used to transform new data sets.
}
