% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load_transform_models}
\alias{load_transform_models}
\title{Load transform models into a cell_data_set.}
\usage{
load_transform_models(cds, directory_path)
}
\arguments{
\item{cds}{a cell_data_set to be transformed using the models.}

\item{directory_path}{a string giving the name of the directory
from which to read the model files.}
}
\value{
a cell_data_set with the transform models loaded by
load_transform_models.
}
\description{
Load transform models, which were saved using save_transform_models,
into a cell_data_set. This function over-writes existing models in
the cell_data_set. For more information read the help information
for save_transform_models.
}
