% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load_worm_embryo}
\alias{load_worm_embryo}
\title{Build a cell_data_set from C. elegans embryo data.}
\usage{
load_worm_embryo()
}
\value{
cds object
}
\description{
Build a cell_data_set from C. elegans embryo data.
}
