% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_cells.R, R/partition_cells.R
\name{louvain_clustering}
\alias{louvain_clustering}
\title{function to run louvain clustering algorithm}
\usage{
louvain_clustering(data, pd, k = 20, weight = F, louvain_iter = 1,
  resolution = NULL, random_seed = 0L, verbose = F, ...)

louvain_clustering(data, pd, k = 20, weight = F, louvain_iter = 1,
  resolution = NULL, random_seed = 0L, verbose = F, ...)
}
\arguments{
\item{data}{low dimensional space used to perform graph clustering}

\item{pd}{the dataframe of the phenotype from the cell dataset (pData(cds))}

\item{k}{number of nearest neighbors used for Louvain clustering}

\item{weight}{whether or not to calculate the weight for each edge in the kNN graph}

\item{louvain_iter}{the number of iteraction for louvain clustering}

\item{resolution}{resolution of clustering result, specifiying the granularity of clusters.
Default to not use resolution and the standard igraph louvain clustering algorithm will be used.}

\item{random_seed}{the seed used by the random number generator in louvain-igraph package}

\item{verbose}{Whether to emit verbose output during dimensionality reduction}

\item{...}{extra arguments used to run louvain_R}

\item{data}{low dimensional space used to perform graph clustering}

\item{pd}{the dataframe of the phenotype from the cell dataset (pData(cds))}

\item{k}{number of nearest neighbors used for Louvain clustering}

\item{weight}{whether or not to calculate the weight for each edge in the kNN graph}

\item{louvain_iter}{the number of iteraction for louvain clustering}

\item{resolution}{resolution of clustering result, specifiying the granularity of clusters.
Default to not use resolution and the standard igraph louvain clustering algorithm will be used.}

\item{random_seed}{the seed used by the random number generator in louvain-igraph package}

\item{verbose}{Whether to emit verbose output during dimensionality reduction}

\item{...}{extra arguments used to run louvain_R}
}
\value{
a list with four elements (g (igraph object for the kNN graph), coord (coordinates of the graph with
layout_component, if the number of cells is less than 3000), edge_links (the data frame to plot the edges of
the igraph, if the number of cells is less than 3000) and optim_res (the louvain clustering result)).

a list with four elements (g (igraph object for the kNN graph), coord (coordinates of the graph with
layout_component, if the number of cells is less than 3000), edge_links (the data frame to plot the edges of
the igraph, if the number of cells is less than 3000) and optim_res (the louvain clustering result)).
}
\description{
function to run louvain clustering algorithm

Function to run louvain clustering algorithm
}
