% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_neighbors.R
\name{make_nn_index}
\alias{make_nn_index}
\title{Make a nearest neighbor index.}
\usage{
make_nn_index(subject_matrix, nn_control = list(), verbose = FALSE)
}
\arguments{
\item{subject_matrix}{the matrix used to build the
index.}

\item{nn_control}{a list of parameters used to make the
nearest neighbor index. See the set_nn_control help
for details.}

\item{verbose}{a boolean indicating whether to emit verbose output.}
}
\value{
a nearest neighbor index.
}
\description{
Make a nearest neighbor index from the
subject_matrix using either the default nearest
neighbor method or the method specified in the
nn_control list parameter. The function returns
the index.
}
\examples{
  \donttest {
    cds <- load_worm_embryo()
    cds <- preprocess_cds(cds)
    nn_index <- make_nn_index(reducedDims(cds)[['PCA']])
  }

}
