% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_genes.R
\name{my.aggregate.Matrix}
\alias{my.aggregate.Matrix}
\title{A function to aggregate columns within a matrix.}
\usage{
my.aggregate.Matrix(x, groupings = NULL, form = NULL, fun = "sum", ...)
}
\description{
A function to aggregate columns within a matrix.
}
\keyword{internal}
