% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalized_counts}
\alias{normalized_counts}
\title{Return a size-factor normalized and (optionally) log-transformed expression matrix}
\usage{
normalized_counts(cds, norm_method = c("log", "binary", "size_only"),
  pseudocount = 1)
}
\description{
Return a size-factor normalized and (optionally) log-transformed expression matrix
}
