% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_cells.R
\name{order_cells}
\alias{order_cells}
\title{Orders cells according to pseudotime.}
\usage{
order_cells(cds, reduction_method = "UMAP", root_pr_nodes = NULL,
  root_cells = NULL, verbose = FALSE)
}
\arguments{
\item{cds}{the cell_data_set upon which to perform this operation}

\item{reduction_method}{a string specifying the reduced dimension method to
use when ordering cells. Currently only "UMAP" is supported.}

\item{root_pr_nodes}{NULL or a vector of starting principal points. If
provided, Pseudotime will start (i.e. be zero) at these graph nodes. Both
\code{root_pr_nodes} and \code{root_cells} cannot be provided.}

\item{root_cells}{Null or a vector of starting cells. If provided,
Pseudotime will start (i.e. be zero) at these cells.}

\item{verbose}{Whether to show running information for order_cells}

\item{reverse}{Whether to reverse the direction of the trajectory}
}
\value{
an updated cell_data_set object, in which colData contains values
for State and Pseudotime for each cell
}
\description{
Learns a "trajectory" describing the biological process the cells are
going through, and calculates where each cell falls within that trajectory.
This function takes as input a cell_data_set and returns it with
two new columns: \code{Pseudotime} and \code{State}, which together encode
where each cell maps to the trajectory. \code{order_cells()} optionally
takes "root" state(s), which you can use to specify the start of the
trajectory. If you don't provide a root state, an plot will be generated
where you can choose the root state(s) interactively. The trajectory will be
composed of segments. The cells from a segment will share the same value of
\code{State}.
}
