% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_cells.R
\name{partition_cells}
\alias{partition_cells}
\title{This function tries to partition cells into different graphs based on a similar approach proposed by Alex Wolf and colleagues}
\usage{
partition_cells(cds, partition_names = NULL, use_pca = FALSE, k = 20,
  weight = F, louvain_iter = 1, resolution = NULL,
  louvain_qval = 0.05, return_all = FALSE, verbose = FALSE, ...)
}
\arguments{
\item{cds}{the CellDataSet upon which to perform this operation}

\item{partition_names}{Which partition groups (column in the pData) should be used to calculate the connectivity between partitions}

\item{use_pca}{Whether or not to cluster cells based on top PCA component. Default to be FALSE.}

\item{k}{number of nearest neighbors used for Louvain clustering (pass to louvain_clustering function)}

\item{weight}{whether or not to calculate the weight for each edge in the kNN graph (pass to louvain_clustering function)}

\item{louvain_iter}{the number of iteraction for louvain clustering (pass to louvain_clustering function)}

\item{resolution}{resolution of clustering result, specifiying the granularity of clusters.
Default to not use resolution and the standard igraph louvain clustering algorithm will be used.}

\item{louvain_qval}{The q-val threshold used to determine the partition of cells (pass to compute_louvain_connected_components)}

\item{return_all}{Whether to return all saved objects from compute_louvain_connected_components function.}

\item{verbose}{Whether to emit verbose output during louvain clustering}

\item{...}{additional arguments to pass to the smoothEmbedding function}
}
\value{
an updated CellDataSet object
}
\description{
Recently Alex Wolf and colleague first proposed the idea to represent the data with an "abstract partition graph"
of clusters identified by Louvain clustering by simply connecting significantly overlapping Louvain clusters (Wolf et al. 2017).
Similar methods for "abstract partition graph" are also recently developed and applied in analyzing the zebrafish / frog cell
atlas datasets (Wagner et al. 2018; Briggs et al. 2018). This coarse-graining representation of the data address a few limitations
of tree-based trajectory inference algorithms, for example, the default principal tree learning algorithm (DDRTree) in Monocle 2.
Although the particion graph doesn't learn an explicit simplified principal tree as DDRTree in Monocle 2, it can naturally separate
outlier cell groups and potentially also parallel trajectories while DDRTree often requires pre-processing before hand to robustly
reconstruct trajectory and cannot handle non-tree like structure. Instead of directly learn a coarse-graining graph of clusters,
we instead take advantage of the participation graph and use it as merely a heuristic initial condition for L1-graph algorithm
(as explained in the next section) to learn the principal points and principal graph at the same time directly from the reduced
UMAP data space. In contrast to the cluster participation method, the principal graph learnt provides an abstraction of the data
manifold while also preserves the local information from the original data space as it is directly embedded in the original data space.
In Monocle 3, we use the clustering_louvain function from the igraph package to perform community detection and implemented an efficient
version of "abstract partition graph" from Alex Wolf. Basically, we first create a design matrix \eqn{X}{} representing the allocation of
each cell to a particular louvain cluster. The column of \eqn{X}{} represents a louvain cluster while the row of \eqn{X}{} a particular cell.
\eqn{X_{ij} = 1}{} if cell \eqn{i}{} belongs to cluster \eqn{j}{}, otherwise 0. We can further obtain the adjacency matrix \eqn{A}{} of the kNN graph
used to perform the louvain clustering where \eqn{A_{ij} = 1}{} if cell \eqn{i}{} connects to \eqn{j}{} in the kNN graph. Then the connection
matrix \eqn{M}{} between each cluster is calculated as, \eqn{M \cdot {X'}{A}{X}}{}. Once \eqn{M}{} is constructed, we can then follow
Supplemental Note 3.1 from (Wolf et al. 2017) to calculate the significance of the connection between each louvain clustering and
consider any clusters with p-value larger than 0.05 by default as not disconnected.
}
