% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{partitions}
\alias{partitions}
\title{Generic to extract partitions from CDS object}
\usage{
partitions(x, reduction_method = "UMAP")
}
\arguments{
\item{x}{A cell_data_set object.}

\item{reduction_method}{Reduced dimension to partitions clusters for.}
}
\description{
Generic to extract partitions from CDS object
}
