% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_cells.R
\name{path_between_nodes}
\alias{path_between_nodes}
\title{traverse from one cell to another cell}
\usage{
path_between_nodes(g, start_pr_nodes, end_pr_nodes)
}
\arguments{
\item{g}{the tree graph learned from monocle 2 during trajectory reconstruction}

\item{starting_cell}{the initial vertex for traversing on the graph}

\item{end_cells}{the terminal vertex for traversing on the graph}
}
\value{
a list of shortest path from the initial cell and terminal cell, geodestic distance between initial cell and terminal cells and branch point passes through the shortest path
}
\description{
traverse from one cell to another cell
}
\keyword{internal}
