% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{pData,cell_data_set-method}
\alias{pData,cell_data_set-method}
\title{Method to access cds colData table}
\usage{
\S4method{pData}{cell_data_set}(x)
}
\arguments{
\item{x}{A cell_data_set object.}
}
\description{
Method to access cds colData table
}
