% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{pData<-,cell_data_set-method}
\alias{pData<-,cell_data_set-method}
\title{Method to set cds colData table}
\usage{
\S4method{pData}{cell_data_set}(x) <- value
}
\arguments{
\item{x}{A cell_data_set object.}

\item{value}{A data frame to set to colData table.}
}
\description{
Method to set cds colData table
}
