% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{pData}
\alias{pData}
\title{Generic to access cds colData table}
\usage{
pData(x)
}
\arguments{
\item{x}{A cell_data_set object.}
}
\value{
colData.
}
\description{
Generic to access cds colData table
}
\examples{
  \donttest{
    cds <- load_a549()
    pData(cds)
  }

}
