% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_3d_cell_trajectory}
\alias{plot_3d_cell_trajectory}
\title{Plot a dataset and trajectory in 3 dimensions}
\usage{
plot_3d_cell_trajectory(cds, downsample_size = NULL, dim = c(1, 2, 3),
  color_by = NULL, palette = NULL, markers = NULL,
  markers_linear = FALSE, cell_size = 5, cell_alpha = 0.5,
  backbone_segment_color = "#77B6EA", backbone_segment_width = 2,
  backbone_vertex_color = NULL, show_group_labels = TRUE,
  show_branch_points = TRUE, webGL_filename = NULL,
  image_filename = NULL, obj_filename = NULL, scale_expr = TRUE,
  width = 800, height = 600, view_matrix = NULL,
  useNULL_GLdev = !interactive(), text_cex = 1, use_plotly = FALSE,
  selfcontained = FALSE, ...)
}
\arguments{
\item{cds}{CellDataSet for the experiment}

\item{dim}{the dimensions used to create the 3D plot, by default it is the first three dimensions}

\item{color_by}{the cell attribute (e.g. the column of pData(cds)) to map to each cell's color}

\item{palette}{the color palette used for plotting}

\item{markers}{a gene name or gene id to use for setting the size of each cell in the plot}

\item{markers_linear}{a boolean used to indicate whether you want to scale the markers logarithimically or linearly}

\item{cell_size}{size of cells, default value is 5}

\item{cell_alpha}{Alpha value for the vertex in the 3D plot}

\item{backbone_segment_color}{color of backbone, value is a string containing a hex color value}

\item{backbone_vertex_color}{Color for the vertex on the principal graph backbone}

\item{show_group_labels}{Whether or not to show labels of groups}

\item{show_branch_points}{Whether to show icons for each branch point (only available after running assign_cell_states)}

\item{webGL_filename}{the name of a file to which you'd to write a webGL file containing the plot}

\item{image_filename}{the name of a file to which you'd to write an animated GIF containing the plot}

\item{obj_filename}{The filename that you'd like to save the 3d plot to}

\item{scale_expr}{whether to tranform the log expression values to z scores}

\item{width}{the width of the plot in pixels}

\item{height}{the height of the plot in pixels}

\item{view_matrix}{A 4x4 matrix signifying your point of view on the 3d plot}

\item{useNULL_GLdev}{if TRUE, don't show the plot on the screen (to be used with webGL or movie output)}

\item{...}{Extra arguments passed to the function}
}
\value{
a ggplot2 plot object
}
\description{
Plot a dataset and trajectory in 3 dimensions
}
\examples{
\dontrun{
plot_3d_cell_trajectory(cds, markers=c("Rbfox3, Neurod1", "Sox2"))
}

}
