% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_cells}
\alias{plot_cells}
\title{Plots the cells along with their trajectories.}
\usage{
plot_cells(cds, x = 1, y = 2, reduction_method = c("UMAP", "tSNE",
  "PCA"), color_cells_by = "cluster", group_cells_by = c("cluster",
  "partition"), genes = NULL, show_trajectory_graph = TRUE,
  trajectory_graph_color = "grey28",
  trajectory_graph_segment_size = 0.75, norm_method = c("log",
  "size_only"), label_cell_groups = TRUE,
  label_groups_by_cluster = TRUE, group_label_size = 2,
  labels_per_group = 1, label_branch_points = TRUE,
  label_roots = TRUE, label_leaves = TRUE, graph_label_size = 2,
  cell_size = 0.35, alpha = 1, min_expr = 0.1, rasterize = FALSE)
}
\arguments{
\item{cds}{cell_data_set for the experiment}

\item{x}{the column of reducedDimS(cds) to plot on the horizontal axis}

\item{y}{the column of reducedDimS(cds) to plot on the vertical axis}

\item{cell_size}{The size of the point for each cell}

\item{alpha}{The alpha aesthetics for the original cell points, useful to
highlight the learned principal graph}

\item{rasterize}{}
}
\value{
a ggplot2 plot object
}
\description{
Plots the cells along with their trajectories.
}
\examples{
\dontrun{
lung <- load_lung()
plot_cells(lung)
plot_cells(lung, color_by="pseudotime", show_backbone=FALSE)
plot_cells(lung, markers="MYH3")
}
}
