% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_cells_per_sample_and_perturbation}
\alias{plot_cells_per_sample_and_perturbation}
\title{Make a boxplot of the number of cells per sample, with each perturbation as a separate box. Optionally facet by a second value.}
\usage{
plot_cells_per_sample_and_perturbation(
  cds,
  perturbation_col = "perturbation",
  count_per_sample_col = "count_per_embryo",
  zscore = FALSE,
  cutoff = 1000,
  facet_by = NULL,
  color_palette = NULL,
  yticks = NULL
)
}
\arguments{
\item{cds}{A cell_data_set for plotting.}

\item{perturbation_col}{How to group the cells for the plot. Must be a column of the colData.}

\item{count_per_sample_col}{Which column of the colData to put in the boxplots.}

\item{zscore}{logical value indicating whether to z-score counts before plotting. If \code{TRUE}, then the values in \code{count_per_sample_col} are grouped by the \code{perturbation_col} and z-scored.}

\item{cutoff}{The minimum number of cells per sample, drawn as a horizontal line. If \code{zscore == TRUE} then this corresponds to the absolute value of the z-score cutoff, and two horizontal lines are drawn.}

\item{facet_by}{Facet the plot by this column of the colData. Each unique value is its own facet.}

\item{color_palette}{List of colors to color perturbation groups. Default is NULL. When NULL, use a default set.}

\item{yticks}{List of numeric values to put on the y-axis ticks.}
}
\value{
A ggplot2 object.
}
\description{
Make a boxplot of the number of cells per sample, with each perturbation as a separate box. Optionally facet by a second value.
}
