% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_mito_umi_per_cell}
\alias{plot_mito_umi_per_cell}
\title{Plot a histogram of the percentage of UMIs mapping to mitochondria for each cell.}
\usage{
plot_mito_umi_per_cell(cds, max_mito_pct = .max_mito_pct, pseudocount = 0.001)
}
\arguments{
\item{cds}{A cell_data_set for plotting.}

\item{max_mito_pct}{Cutoff for the maximum percentage of UMIs mapping to mitochondria.}

\item{pseudocount}{Value to add to the percentages before taking the log. Note that this is in terms of percent, not fractional.}
}
\value{
A ggplot2 object.
}
\description{
Plot a histogram of the percentage of UMIs mapping to mitochondria for each cell.
}
