% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_pc_variance_explained}
\alias{plot_pc_variance_explained}
\title{Plots the fraction of variance explained by the each component based on
PCA from the normalized expression data determined using preprocess_cds.
This is the fraction of the component variance relative to the variance
of the components retained in the PCA; not the total variance.}
\usage{
plot_pc_variance_explained(cds)
}
\arguments{
\item{cds}{cell_data_set of the experiment.}
}
\value{
ggplot object.
}
\description{
Plots the fraction of variance explained by the each component based on
PCA from the normalized expression data determined using preprocess_cds.
This is the fraction of the component variance relative to the variance
of the components retained in the PCA; not the total variance.
}
\examples{
  \donttest{
    cds <- load_a549()
    cds <- preprocess_cds(cds)
    plot_pc_variance_explained(cds)
  }

}
