% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_percent_cells_positive}
\alias{plot_percent_cells_positive}
\title{Plots the number of cells expressing one or more genes above a given value
as a barplot}
\usage{
plot_percent_cells_positive(cds_subset, group_cells_by = NULL,
  min_expr = NULL, nrow = NULL, ncol = 1, panel_order = NULL,
  plot_as_count = FALSE, label_by_short_name = TRUE,
  normalize = TRUE, plot_limits = NULL)
}
\arguments{
\item{cds_subset}{Subset cell_data_set to be plotted.}

\item{min_expr}{the minimum (untransformed) expression level to consider the
gene 'expressed'. If \code{NULL},
zero is used.}

\item{nrow}{the number of panels per row in the figure.}

\item{ncol}{the number of panels per column in the figure.}

\item{panel_order}{the order in which genes should be layed out
(left-to-right, top-to-bottom)}

\item{plot_as_count}{whether to plot as a count of cells rather than a
percent.}

\item{label_by_short_name}{label figure panels by gene_short_name (TRUE) or
feature id (FALSE).}

\item{plot_limits}{A pair of number specifying the limits of the y axis. If
\code{NULL}, scale to the range of the data. Example \code{c(0,100)}.}

\item{grouping}{the cell attribute (e.g. the column of colData(cds)) to
group cells by on the horizontal axis}
}
\value{
a ggplot2 plot object
}
\description{
@description Accepts a cell_data_set and the parameter "grouping", used for
dividing cells into groups. Returns one or more bar graphs (one graph for
each gene in the cell_data_set). Each graph shows the percentage of cells
that express a gene in each sub-group in the cell_data_set.
}
\details{
As an example, let's say the cell_data_set passed into the function as
cds_subset included genes A, B, and C and the grouping parameter divided
the cells into three groups called X, Y, and Z. Then three graphs would be
produced called A, B, and C. In each graph there would be three bars one
for X, one for Y, and one for Z. The X bar in the A graph would show the
percentage of cells in the X group that express gene A.
}
\examples{
\dontrun{
library(HSMMSingleCell)
HSMM <- load_HSMM()
MYOG_ID1 <- HSMM[row.names(subset(rowData(HSMM),
                                  gene_short_name \%in\% c("MYOG", "ID1"))),]
plot_percent_cells_positive(MYOG_ID1, grouping="Media", ncol=2)
}
}
