% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{principal_graph,cell_data_set-method}
\alias{principal_graph,cell_data_set-method}
\title{Method to extract principal graph from CDS}
\usage{
\S4method{principal_graph}{cell_data_set}(x)
}
\arguments{
\item{x}{A cell_data_set object.}
}
\value{
Principle graph.
}
\description{
Method to extract principal graph from CDS
}
