% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{principal_graph<-,cell_data_set-method}
\alias{principal_graph<-,cell_data_set-method}
\title{Generic to set principal graph to CDS}
\usage{
\S4method{principal_graph}{cell_data_set}(x) <- value
}
\arguments{
\item{x}{A cell_data_set object.}

\item{value}{A principal graph object.}
}
\description{
Generic to set principal graph to CDS
}
