% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-cell_data_set.R
\name{principal_graph<-}
\alias{principal_graph<-}
\title{Set the minimum spanning tree generated by Monocle during cell ordering.}
\usage{
principal_graph(cds) <- value
}
\arguments{
\item{cds}{A CellDataSet object.}

\item{value}{an igraph object describing the minimum spanning tree.}
}
\value{
An updated CellDataSet object
}
\description{
Sets the minimum spanning tree used by Monocle during cell ordering. Not intended to be called directly.
}
\examples{
\dontrun{
cds <- principal_graph(T)
}
}
