% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{principal_graph}
\alias{principal_graph}
\title{Generic to extract principal graph from CDS}
\usage{
principal_graph(x)
}
\arguments{
\item{x}{A cell_data_set object.}
}
\description{
Generic to extract principal graph from CDS
}
