% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{principal_graph_aux,cell_data_set-method}
\alias{principal_graph_aux,cell_data_set-method}
\title{Method to extract principal graph auxiliary information from CDS}
\usage{
\S4method{principal_graph_aux}{cell_data_set}(x)
}
\arguments{
\item{x}{A cell_data_set object.}
}
\value{
Principal graph auxiliary information.
}
\description{
Method to extract principal graph auxiliary information from CDS
}
