% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{principal_graph_aux<-}
\alias{principal_graph_aux<-}
\title{Generic to set principal graph auxilliary information into CDS}
\usage{
principal_graph_aux(x) <- value
}
\arguments{
\item{x}{A cell_data_set object.}

\item{value}{A SimpleList of principal graph auxilliary information.}
}
\description{
Generic to set principal graph auxilliary information into CDS
}
