% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_graph.R
\name{pruneTree_in_learnGraph}
\alias{pruneTree_in_learnGraph}
\title{Function to prune the graph}
\usage{
pruneTree_in_learnGraph(stree_ori, stree_loop_clousre,
  minimal_branch_len = 10)
}
\description{
Function to prune the graph
}
