% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{pseudotime,cell_data_set-method}
\alias{pseudotime,cell_data_set-method}
\title{Method to extract pseudotime from CDS object}
\usage{
\S4method{pseudotime}{cell_data_set}(x, reduction_method = "UMAP")
}
\arguments{
\item{cell_data_set}{A cell_data_set object.}
}
\description{
Method to extract pseudotime from CDS object
}
