% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{pseudotime}
\alias{pseudotime}
\title{Generic to extract pseudotime from CDS object}
\usage{
pseudotime(x, reduction_method = "UMAP")
}
\arguments{
\item{x}{CDS object}

\item{reduction_method}{Reduced dimension to extract pseudotime for.}
}
\description{
Generic to extract pseudotime from CDS object
}
