% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-cell_data_set.R
\name{reducedDimA<-}
\alias{reducedDimA<-}
\title{Get the weights needed to lift cells back to high dimensional expression space.}
\usage{
reducedDimA(cds) <- value
}
\arguments{
\item{cds}{A cell_data_set object.}

\item{value}{A whitened expression data matrix}
}
\value{
An updated cell_data_set object
}
\description{
Sets the weights transform the cells' coordinates in the reduced dimension
space back to the full (whitened) space.
}
\examples{
\dontrun{
cds <- reducedDimA(A)
}
}
