% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-cell_data_set.R
\name{reducedDimA}
\alias{reducedDimA}
\title{Get the weights needed to lift cells back to high dimensional expression space.}
\usage{
reducedDimA(cds)
}
\arguments{
\item{cds}{A cell_data_set object.}
}
\value{
A matrix that when multiplied by a reduced-dimension set of coordinates for the cell_data_set,
recovers a matrix in the full (whitened) space
}
\description{
Retrieves the weights that transform the cells' coordinates in the reduced
dimension space back to the full (whitened) space.
}
\examples{
\dontrun{
A <- reducedDimA(HSMM)
}
}
