% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-cell_data_set.R
\docType{methods}
\name{reducedDimK<-}
\alias{reducedDimK<-}
\title{Sets the the whitening matrix during independent component analysis.}
\usage{
reducedDimK(cds) <- value
}
\arguments{
\item{cds}{A cell_data_set object.}

\item{value}{a numeric matrix}
}
\value{
A matrix, where each row is a set of whitened expression values for a feature and columns are cells.
}
\description{
Sets the the whitening matrix during independent component analysis.
}
\examples{
\dontrun{
cds <- reducedDimK(K)
}
}
