% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-cell_data_set.R
\docType{methods}
\name{reducedDimK}
\alias{reducedDimK}
\title{Retrieves the the whitening matrix during independent component analysis.}
\usage{
reducedDimK(cds)
}
\arguments{
\item{cds}{A cell_data_set object.}
}
\value{
A matrix, where each row is a set of whitened expression values for a feature and columns are cells.
}
\description{
Retrieves the the whitening matrix during independent component analysis.
}
\examples{
\dontrun{
K <- reducedDimW(HSMM)
}
}
