% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-cell_data_set.R
\name{reducedDimS<-}
\alias{reducedDimS<-}
\title{Set embedding coordinates of each cell in a cell_data_set}
\usage{
reducedDimS(cds) <- value
}
\arguments{
\item{cds}{A CellDataSet object.}

\item{value}{A matrix of coordinates specifying each cell's position in the reduced-dimensionality space.}
}
\value{
An update CellDataSet object
}
\description{
This function sets the coordinates of each cell in a new
(reduced-dimensionality) space. Not intended to be called directly.
}
\examples{
\dontrun{
cds <- reducedDimS(S)
}
}
