% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-cell_data_set.R
\name{reducedDimS}
\alias{reducedDimS}
\title{Retrieves the coordinates of each cell in the reduced-dimensionality space generated by calls to
reduceDimension.}
\usage{
reducedDimS(cds)
}
\arguments{
\item{cds}{A cell_data_set object.}
}
\value{
A matrix, where rows are cell coordinates and columns correspond to dimensions of the
reduced space.
}
\description{
Reducing the dimensionality of the expression data is a core step in the Monocle
workflow. After you call reduceDimension(), this function will return the new
coordinates of your cells in the reduced space.
}
\examples{
\dontrun{
S <- reducedDimS(HSMM)
}
}
