% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_graph.R
\name{repmat}
\alias{repmat}
\title{function to reproduce the behavior of repmat function in matlab to replicate and tile an matrix}
\usage{
repmat(X, m, n)
}
\arguments{
\item{X}{matrix for tiling and replicate the data}

\item{m}{a numeric value for tiling a matrix}

\item{n}{a numeric value for tiling a matrix}
}
\value{
a matrix
}
\description{
function to reproduce the behavior of repmat function in matlab to replicate and tile an matrix
}
