% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_cells.R
\name{root_nodes}
\alias{root_nodes}
\title{Return the names of principal graph nodes that are roots}
\usage{
root_nodes(cds, reduction_method = "UMAP")
}
\description{
Return the names of principal graph nodes that are roots
}
