% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_align_cds_model}
\alias{save_align_cds_model}
\title{Save align_cds model.}
\usage{
save_align_cds_model(
  cds,
  method = c("Aligned"),
  file_name_root = NULL,
  comment = NULL
)
}
\arguments{
\item{cds}{cell_data_set with an existing model, which was created using
cds <- preprocess_cds(cds, build_nn_index=TRUE).}

\item{comment}{An optional string that describes the model.}

\item{file_name_root.}{A string with the root of names given to the
files to which the model will be saved. The files are named
<file_name_root>.aln_rds and <file_name_root>.aln_nn_index<n>. Note: the
<file_name_root>.aln_rds file contains only the information needed to
load the align_cds model into the cds object given in
save_align_cds_model(cds,...).}
}
\value{
None
}
\description{
Save an align_cds model consisting of dimensionality reduction
models and align_cds nearest neighbor index objects, which are
used to transform new data sets. Additionally, it saves the
preprocess_cds nearest neighbor index, if preprocess_cds() was run with
build_nn_index=TRUE.
}
