% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retired.R
\name{save_monocle_objects_retired}
\alias{save_monocle_objects_retired}
\title{Save a Monocle3 full cell_data_set.}
\usage{
save_monocle_objects_retired(cds, directory_path)
}
\arguments{
\item{cds}{cell_data_set to save.}

\item{directory_path}{A string giving the directory in
which to write the files.}
}
\value{
None.
}
\description{
Save a Monocle3 full cell_data_set to a specified directory
by writing the R objects to an RDS file and
the Annoy nearest neighbor indexes to individual index files.
}
