% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_preprocess_classifier}
\alias{save_preprocess_classifier}
\title{Save preprocess classifier.}
\usage{
save_preprocess_classifier(
  cds,
  method = c("PCA", "LSI"),
  file_root_name = NULL,
  comment = NULL
)
}
\arguments{
\item{cds}{cell_data_set with existing classifier, which was created using
cds <- preprocess_cds(cds, build_nn_index=TRUE).}

\item{method}{A string indicating the method used to build the classifier
that you want saved. Default is "PCA".}

\item{comment}{An optional string that describes the classifier.}

\item{file_root_name.}{A string giving the root of names given to the two
files to which the classifier is saved. The files are named <file_root_name>.rds
and <file_root_name>.annoy_index.}
}
\value{
None
}
\description{
Save a preprocess classifier consisting of dimensionality reduction
model and Annoy nearest neighbor data structures, which are used to
transform and classify new data sets.
}
