% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_preprocess_model}
\alias{save_preprocess_model}
\title{Save preprocess model.}
\usage{
save_preprocess_model(
  cds,
  method = c("PCA", "LSI"),
  file_root_name = NULL,
  comment = NULL
)
}
\arguments{
\item{cds}{cell_data_set with existing model, which was created using
cds <- preprocess_cds(cds, build_nn_index=TRUE).}

\item{method}{A string indicating the method used to build the model
that you want saved. Default is "PCA".}

\item{comment}{An optional string that describes the model.}

\item{file_root_name.}{A string giving the root of names given to the two
files to which the model is saved. The files are named <file_root_name>.rds
and <file_root_name>.ppc_nn_index.}
}
\value{
None
}
\description{
Save a preprocess model consisting of dimensionality reduction
model and Annoy nearest neighbor index objects, which are used to
transform new data sets.
}
