% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retired.R
\name{save_preprocess_model_retired}
\alias{save_preprocess_model_retired}
\title{Save preprocess model.}
\usage{
save_preprocess_model_retired(
  cds,
  method = c("PCA", "LSI"),
  file_name_root = NULL,
  comment = NULL
)
}
\arguments{
\item{cds}{cell_data_set with an existing model, which was created using
cds <- preprocess_cds(cds, build_nn_index=TRUE).}

\item{method}{A string indicating the method used to build the model
that you want to save. Default is "PCA".}

\item{comment}{An optional string that describes the model.}

\item{file_name_root.}{A string with the root of names given to the two
files to which the model will be saved. The files are named
<file_name_root>.ppc_rds and <file_name_root>.ppc_nn_index. Note: the
<file_name_root>.ppc_rds file contains only the information needed to load
the preprocess model into the cds object given in
load_preprocess_model(cds,...).}
}
\value{
None
}
\description{
Save a preprocess model consisting of dimensionality reduction
model and preprocess nearest neighbor index objects, which
are used to transform new data sets.
}
