% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_reduce_dimension_classifier}
\alias{save_reduce_dimension_classifier}
\title{Save reduce_dimension classifier.}
\usage{
save_reduce_dimension_classifier(
  cds,
  preprocess_method = c("PCA"),
  method = c("UMAP"),
  file_root_name = NULL,
  comment = NULL
)
}
\arguments{
\item{cds}{cell_data_set with existing classifier, which was created using
cds <- reduce_dimension(cds, build_nn_index=TRUE).}

\item{method}{A string indicating the method used to build the classifier
that you want saved. This must be "UMAP". Default = "UMAP".}

\item{comment}{An optional string that describes the classifier.}

\item{file_root_name.}{A string giving the root of names given to the
two files to which the classifier is saved. The files are named
<file_root_name>.rds, and <file_root_name>.annoy_index.}
}
\value{
None
}
\description{
Save a reduce_dimension classifier consisting of dimensionality reduction
models and Annoy nearest neighbor data structures, which are used to transform and
classify new data sets.
}
