% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_reduce_dimension_model}
\alias{save_reduce_dimension_model}
\title{Save reduce_dimension model.}
\usage{
save_reduce_dimension_model(
  cds,
  method = c("UMAP"),
  file_root_name = NULL,
  comment = NULL
)
}
\arguments{
\item{cds}{cell_data_set with existing model, which was created using
cds <- reduce_dimension(cds, build_nn_index=TRUE).}

\item{method}{A string indicating the method used to build the model
that you want saved. This must be "UMAP". Default = "UMAP".}

\item{comment}{An optional string that describes the model.}

\item{file_root_name.}{A string giving the root of names given to the
two files to which the model is saved. The files are named
<file_root_name>.rds, <file_root_name>._umap_nn_index, and
<file_root_name>._rdd_nn_index.}
}
\value{
None
}
\description{
Save a reduce_dimension model consisting of dimensionality reduction
models and Annoy nearest neighbor index objects, which are used to
transform and classify new data sets.
}
