% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retired.R
\name{save_reduce_dimension_model_retired}
\alias{save_reduce_dimension_model_retired}
\title{Save reduce_dimension model.}
\usage{
save_reduce_dimension_model_retired(
  cds,
  method = c("UMAP"),
  file_name_root = NULL,
  comment = NULL
)
}
\arguments{
\item{cds}{cell_data_set with an existing model, which was created using
cds <- reduce_dimension(cds, build_nn_index=TRUE).}

\item{method}{A string indicating the method used to build the model
that you want saved. This must be "UMAP". Default = "UMAP".}

\item{comment}{An optional string that describes the model.}

\item{file_name_root.}{A string with the root of names given to the
files to which the model will be saved. The files are named
<file_name_root>.rdd_rds, <file_name_root>.rdd_umap_nn_index, and
<file_name_root>._rdd_nn_index<n>. Note: the <file_name_root>.rdd_rds file
contains only the information needed to load the reduce_dimension
model into the cds object given in load_reduce_dimension_model(cds,...).}
}
\value{
None
}
\description{
Save a reduce_dimension model consisting of dimensionality
reduction models and reduce_dimension nearest neighbor
index objects, which are used to transform new data
sets. Additionally, it saves preprocess and align_cds nearest neighbor
indexes, when preprocess_cds() and align_cds() were run with
build_nn_index=TRUE.
}
