% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_neighbors.R
\name{search_nn_matrix}
\alias{search_nn_matrix}
\title{Search a subject matrix for nearest neighbors to a query_matrix.}
\usage{
search_nn_matrix(
  subject_matrix,
  query_matrix,
  k = 25,
  nn_control = list(),
  verbose = FALSE
)
}
\arguments{
\item{subject_matrix}{a matrix used to build a nearest neighbor index.}

\item{query_matrix}{a matrix used to search the subject_matrix
nearest neighbor index.}

\item{k}{an integer for the number of nearest neighbors to return for
each cell. Default is 25.}

\item{nn_control}{a list of parameters used to make and search the nearest
neighbor index. See the set_nn_control help for details.}

\item{verbose}{a boolean indicating whether to emit verbose output.}
}
\value{
a list list(nn.idx, nn.dists) where nn.idx is
a matrix of nearest neighbor indices and nn.dists is a matrix of
the distance between the index given by the row number and the index
given in nn.idx. If the query_matrix is the same as the subject
matrix, the index given by the row number should be in the row, usually
in the first column.
}
\description{
Make a nearest neighbors index using the subject matrix and
search it for nearest neighbors to the query_matrix.
}
