% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_neighbors.R
\name{set_cds_nn_index}
\alias{set_cds_nn_index}
\title{Set a nearest neighor index in the cell_data_set.}
\usage{
set_cds_nn_index(
  cds,
  reduction_method = c("UMAP", "PCA", "LSI", "Aligned", "tSNE"),
  nn_index,
  verbose = FALSE
)
}
\arguments{
\item{cds}{a cell_data_set in which to store the nearest
neighbor index.}

\item{reduction_method}{a string giving the reduced
dimension matrix used to make the nn_index nearest
neighbor index, and determines where the index is
stored in the cell_data_set.}

\item{nn_index}{a nearest neighbor index to store in cds.}

\item{verbose}{a boolean indicating whether to emit verbose output.}
}
\value{
a cell_data_set with the stored index.
}
\description{
Store the given nearest neighbor index
in the cell_data_set. The reduction_method parameter
tells set_cds_nn_index where in the cell_data_set to
store the index.
}
