% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_neighbors.R
\name{set_cds_nn_search}
\alias{set_cds_nn_search}
\title{on this matrix so call this function immediately after running the search.}
\usage{
set_cds_nn_search(
  cds,
  reduction_method = c("UMAP", "PCA", "LSI", "Aligned", "tSNE"),
  search_id,
  k,
  nn_control = list(),
  verbose = TRUE
)
}
\description{
on this matrix so call this function immediately after running the search.
}
