% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_ordering_filter}
\alias{set_ordering_filter}
\title{Marks genes for clustering}
\usage{
set_ordering_filter(cds, ordering_genes)
}
\arguments{
\item{cds}{the cell_data_set upon which to perform this operation}

\item{ordering_genes}{a vector of feature ids (from the cell_data_set's featureData) used for ordering cells}
}
\value{
an updated cell_data_set object
}
\description{
The function marks genes that will be used for clustering in subsequent calls to clusterCells.
The list of selected genes can be altered at any time.
}
