% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{set_pca_control}
\alias{set_pca_control}
\title{Verify and set the pca_control parameter list.}
\usage{
set_pca_control(
  pca_control = list(),
  pca_control_default = list(),
  assay_control = list()
)
}
\arguments{
\item{pca_control}{Input control list.}

\item{pca_control_default}{Ignored at this time.}

\item{assay_control}{The assay_control list used to store the
counts matrix. This is stored in
metadata(assays(cds)[\link{'counts'}][\link{'assay_control'}].}
}
\value{
pca_control Output control list.
}
\description{
Verifies and sets the list of parameter values
that is used when storing matrices that are used for PCA
and LSI. To see the default values,
call "set_pca_control(pca_control=list(show_values=TRUE))".
"show_values=TRUE" can be used in functions with the
pca_control list parameter in which case the function will
show the pca_control values to be used and then stop.
}
\section{pca_control parameters}{

\describe{
\item{matrix_class}{Specifies the matrix class to use for
matrix storage. The acceptable values are "dgCMatrix"
and "BPCells".}
\item{matrix_type}{Specifies whether to store the matrix
single precision "floats" (matrix_type="float") or
double precision "doubles" (matrix_type="double").
"matrix_type" is used only for BPCells class matrices.}
\item{matrix_mode}{Specifies whether to store the BPCells
class matrix in memory (matrix_mode="mem") or on
disk (matrix_mode="dir"). "matrix_mode" is used only
for BPCells class matrices.}
\item{matrix_path}{Specifies the directory where the
BPCells on-disk matrix data are stored in a
sub-directory with a randomized name. The default is
in the directory where R is running. "matrix_path" is
used only for BPCells class matrices with
matrix_mode="dir". For example, if matrix_path is set
to "/tmp" Monocle3 will create a directory with a
name that has the form "monocle.bpcells.*.tmp" in
"/tmp". The asterisk represents a random string that
makes the names unique.}
\item{matrix_compress}{Specifies whether to use bit-packing
compression to store BPCells matrix values. This
reduces somewhat storage requirements but increases
run time. "matrix_compress" is used only for BPCells
class matrices.}
\item{matrix_buffer_size}{Specifies how many items of
data to buffer in memory before flushing to disk. This
is used for matrix_class="BPCells" with matrix_mode="dir".}
\item{Note: the default values are the same as those specified for
the assay matrix except when assay_control[\link{'matrix_type'}] is
'uint32_t' in which case it is set to 'double' or when
assay_control[\link{'matrix_compress'}] is TRUE in which case
it is set to FALSE.}
}

