% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_cds.R
\name{seurat_to_cds}
\alias{seurat_to_cds}
\title{Import a seurat object and convert it to a monocle cds.}
\usage{
seurat_to_cds(seurat_cds, import_all = FALSE)
}
\arguments{
\item{seurat_cds}{the Seurat object you would like to convert into a monocle3 cds}

\item{import_all}{logical, whether to import all the slots in seurat.
Default is FALSE (or only keep minimal dataset).}
}
\value{
a new monocle3 cell_data_set
}
\description{
This function takes a Seurat object and converts it to a monocle3 cds.
}
