% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-cell_data_set.R
\name{size_factors}
\alias{size_factors}
\title{Set the minimum spanning tree generated by Monocle during cell ordering.}
\usage{
size_factors(cds, value)
}
\arguments{
\item{cds}{A cell_data_set object.}

\item{value}{an igraph object describing the minimum spanning tree.}
}
\value{
An updated cell_data_set object
}
\description{
Sets the minimum spanning tree used by Monocle during cell ordering. Not intended to be called directly.
}
\examples{
\dontrun{
cds <- size_factors(T)
}
}
