% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_genes.R
\name{aggregate_gene_expression}
\alias{aggregate_gene_expression}
\title{Creates a matrix with aggregated expression values for arbitrary groups of
genes}
\usage{
aggregate_gene_expression(
  cds,
  gene_group_df = NULL,
  cell_group_df = NULL,
  norm_method = c("log", "binary", "size_only"),
  pseudocount = 1,
  scale_agg_values = TRUE,
  max_agg_value = 3,
  min_agg_value = -3,
  exclude.na = TRUE,
  gene_agg_fun = "sum",
  cell_agg_fun = "mean"
)
}
\arguments{
\item{cds}{The cell_data_set on which this function operates}

\item{gene_group_df}{A dataframe in which the first column contains gene ids
or short gene names and the second contains groups. If NULL, genes are not
grouped.}

\item{cell_group_df}{A dataframe in which the first column contains cell ids
and the second contains groups. If NULL, cells are not grouped.}

\item{norm_method}{How to transform gene expression values before
aggregating them. If "log", a pseudocount is added. If "size_only", values
are divided by cell size factors prior to aggregation.}

\item{pseudocount}{Value to add to expression prior to log transformation
and aggregation.}

\item{scale_agg_values}{Whether to center and scale aggregated groups of
genes.}

\item{max_agg_value}{If scale_agg_values is TRUE, the maximum value the
resulting Z scores can take. Higher values are capped at this threshold.}

\item{min_agg_value}{If scale_agg_values is TRUE, the minimum value the
resulting Z scores can take. Lower values are capped at this threshold.}

\item{exclude.na}{Logical indicating whether or not to exclude NA values
from the aggregated matrix.}

\item{gene_agg_fun}{Function used for gene aggregation. This
can be either sum or mean. Default is sum.}

\item{cell_agg_fun}{Function used for cell aggregation. Default is mean.}
}
\value{
A matrix of dimension NxM, where N is the number of gene groups and
M is the number of cell groups.
}
\description{
Creates a matrix with aggregated expression values for arbitrary groups of
genes
}
