% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{align_transform}
\alias{align_transform}
\title{Apply an align_cds transform model to a cell_data_set.}
\usage{
align_transform(cds, method = c("Aligned"))
}
\arguments{
\item{cds}{A cell_data_set to be transformed.}

\item{method}{A previously loaded transform model that
is used to reduce the dimensions of the preprocessed
count matrix in cds.}
}
\value{
A cell_data_set with an align_cds transformed
reduced count matrix.
}
\description{
Applies a previously calculated align_cds transform model
to a new preprocess transformed matrix. For more
information read the help information for
save_transform_models.
}
\details{
Note: this function is a place holder. It does not
map the transformed count matrix to aligned space
at this time because I don't know how to do so.
}
