% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_graph.R
\name{calc_principal_graph}
\alias{calc_principal_graph}
\title{Function to automatically learn the structure of data by either using
L1-graph or the spanning-tree formulization}
\usage{
calc_principal_graph(X, C0, maxiter = 10, eps = 1e-05,
  L1.gamma = 0.5, L1.sigma = 0.01, verbose = T)
}
\arguments{
\item{X}{the input data DxN}

\item{C0}{the initialization of centroids}

\item{maxiter}{maximum number of iteration}

\item{eps}{relative objective difference}

\item{L1.gamma}{regularization parameter for k-means (the prefix of 'param'
is used to avoid name collision with gamma)}

\item{L1.sigma}{bandwidth parameter}

\item{verbose}{emit results from iteration}
}
\value{
a list of X, C, W, P, objs
X is the input data
C is the centers for principal graph
W is the principal graph matrix
P is the cluster assignment matrix
objs is the objective value for the function
}
\description{
Function to automatically learn the structure of data by either using
L1-graph or the spanning-tree formulization
}
