% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-cell_data_set.R
\docType{methods}
\name{cell_data_set-methods}
\alias{cell_data_set-methods}
\alias{[,cell_data_set,ANY,ANY,ANY-method}
\alias{cell_data_set,ANY,ANY,ANY-method}
\alias{sizeFactors,cell_data_set-method}
\alias{cell_data_set,ANY,ANY-method}
\alias{sizeFactors<-,cell_data_set,numeric-method}
\alias{estimateSizeFactors,cell_data_set-method}
\alias{estimateDispersions,cell_data_set-method}
\title{Methods for the cell_data_set class}
\usage{
\S4method{[}{cell_data_set,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)

\S4method{sizeFactors}{cell_data_set}(object)

\S4method{sizeFactors}{cell_data_set,numeric}(object) <- value

\S4method{estimateSizeFactors}{cell_data_set}(object, locfunc = median,
  ...)

\S4method{estimateDispersions}{cell_data_set}(object,
  modelFormulaStr = "~ 1", relative_expr = TRUE,
  min_cells_detected = 1, remove_outliers = TRUE, cores = 1, ...)
}
\arguments{
\item{x}{the cell_data_set object}

\item{i}{index (or name) to extract or replace}

\item{j}{index (or name) to extract or replace}

\item{...}{extra arguments passed to method}

\item{drop}{If TRUE the result is coerced to the lowest possible dimension
(see the examples). This only works for extracting elements, not for the
replacement.}

\item{object}{The cell_data_set object}

\item{value}{A vector of size factors, with length equal to the cells in
object}

\item{locfunc}{A function applied to the geometric-mean-scaled expression
values to derive the size factor.}

\item{modelFormulaStr}{A model formula, passed as a string, specifying how
to group the cells prior to estimated dispersion. The default groups all
cells together.}

\item{relative_expr}{Whether to transform expression into relative values}

\item{min_cells_detected}{Only include genes detected above
lower_detection_limit in at least this many cells in the dispersion
calculation}

\item{remove_outliers}{Whether to remove outliers (using Cook's distance)
when estimating dispersions}

\item{cores}{The number of cores to use for computing dispersions}
}
\description{
Methods for the cell_data_set class
}
