% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-cell_data_set.R
\docType{methods}
\name{cell_data_set-methods}
\alias{cell_data_set-methods}
\title{Methods for the cell_data_set class}
\arguments{
\item{object}{The cell_data_set object}
}
\description{
Methods for the cell_data_set class
}
