% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_data_set.R
\docType{class}
\name{cell_data_set}
\alias{cell_data_set}
\alias{cell_data_set-class}
\title{The cell_data_set class}
\description{
The main class used by Monocle3 to hold single-cell expression data.
cell_data_set extends the Bioconductor SingleCellExperiment class.
}
\details{
This class is initialized from a matrix of expression values along with cell
and feature metadata.
}
\section{Fields}{

\describe{
\item{\code{reduce_dim_aux}}{SimpleList, auxiliary information from reduced
dimension.}

\item{\code{principal_graph_aux}}{SimpleList, auxiliary information from principal
graph construction}

\item{\code{principal_graph}}{SimpleList of igraph objects containing principal
graphs for different dimensionality reduction.}

\item{\code{clusters}}{SimpleList of cluster information for different
dimensionality reduction.}
}}

