% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_cells.R
\name{choose_graph_segments}
\alias{choose_graph_segments}
\title{Choose cells along the path of a principal graph}
\usage{
choose_graph_segments(
  cds,
  reduction_method = "UMAP",
  starting_pr_node = NULL,
  ending_pr_nodes = NULL,
  return_list = FALSE,
  clear_cds = TRUE
)
}
\arguments{
\item{cds}{CDS object to be subsetted.}

\item{reduction_method}{The reduction method to plot while choosing cells.
Currently only "UMAP" is supported.}

\item{starting_pr_node}{NULL, or a string with the name of the starting
principal node to be used. You can see the principal nodes in your dataset
by using plot_cells with label_principal_points = TRUE.}

\item{ending_pr_nodes}{NULL, or one or more strings with the name(s) of the
ending principal node(s) to be used. You can see the principal nodes in
your dataset by using plot_cells with label_principal_points = TRUE.}

\item{return_list}{Logical, return a list of cells instead of a subsetted
CDS object.}

\item{clear_cds}{Logical, clear CDS slots before returning.
After clearing the cds, re-run processing from preprocess_cds(), ...
Default is TRUE.}
}
\value{
A subset CDS object. If return_list = FALSE, a list of cell and
graph node names.
}
\description{
Choose cells along the path of a principal graph
}
\examples{
  cds <- load_worm_embryo()
  cds <- preprocess_cds(cds, num_dim=50)
  cds <- align_cds(cds, alignment_group = "batch", residual_model_formula_str = "~ bg.300.loading + bg.400.loading + bg.500.1.loading + bg.500.2.loading + bg.r17.loading + bg.b01.loading + bg.b02.loading")
  cds <- reduce_dimension(cds)
  cds <- cluster_cells(cds)
  cds <- learn_graph(cds)
  cds_choose <- choose_graph_segments(cds, "UMAP")

}
