% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_cells.R
\name{choose_graph_segments}
\alias{choose_graph_segments}
\title{Choose cells interactively along the path of a principal graph}
\usage{
choose_graph_segments(
  cds,
  reduction_method = "UMAP",
  return_list = FALSE,
  clear_cds = TRUE
)
}
\arguments{
\item{cds}{CDS object to be subsetted.}

\item{reduction_method}{The reduction method to plot while choosing cells.
Currently only "UMAP" is supported.}

\item{return_list}{Logical, return a list of cells instead of a subsetted
CDS object.}

\item{clear_cds}{Logical, clear CDS slots before returning.
After clearing the cds, re-run processing from preprocess_cds(), ...
Default is TRUE.}
}
\value{
A subset CDS object. If return_list = FALSE, a list of cell and
graph node names.
}
\description{
Choose cells interactively along the path of a principal graph
}
