% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_models.R
\name{coefficient_table}
\alias{coefficient_table}
\title{Extracts a table of coefficients from a tibble containing model objects}
\usage{
coefficient_table(model_tbl)
}
\description{
Extracts a table of coefficients from a tibble containing model objects
}
