% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine_cds}
\alias{combine_cds}
\title{Combine a list of cell_data_set objects}
\usage{
combine_cds(
  cds_list,
  keep_all_genes = TRUE,
  cell_names_unique = FALSE,
  sample_col_name = "sample"
)
}
\arguments{
\item{cds_list}{List of cds objects to be combined.}

\item{keep_all_genes}{Logical indicating what to do if there is a mismatch
in the gene sets of the CDSs. If TRUE, all genes are kept and cells from
CDSs missing a given gene will be filled in with zeroes. If FALSE, only
the genes in common among all of the CDSs will be kept. Default is TRUE.}

\item{cell_names_unique}{Logical indicating whether all of the cell IDs
across all of the CDSs are unique. If FALSE, the CDS name is appended to
each cell ID to prevent collisions. Default is FALSE.}

\item{sample_col_name}{A string to be the column name for the colData column
that indicates which original cds the cell derives from. Default is
"sample".}
}
\value{
A combined cell_data_set object.
}
\description{
This function will combine a list of cell_data_set objects into a new
cell_data_set object.
}
