% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine_cds}
\alias{combine_cds}
\title{Combine a list of cell_data_set objects}
\usage{
combine_cds(
  cds_list,
  keep_all_genes = TRUE,
  cell_names_unique = FALSE,
  sample_col_name = "sample",
  keep_reduced_dims = FALSE,
  matrix_control = list()
)
}
\arguments{
\item{cds_list}{List of cds objects to be combined.}

\item{keep_all_genes}{Logical indicating what to do if there is a mismatch
in the gene sets of the CDSs. If TRUE, all genes are kept and cells from
CDSs missing a given gene will be filled in with zeroes. If FALSE, only
the genes in common among all of the CDSs will be kept. Default is TRUE.}

\item{cell_names_unique}{Logical indicating whether all of the cell IDs
across all of the CDSs are unique. If FALSE, the CDS name is appended to
each cell ID to prevent collisions. These cell IDs are used as count matrix
column names and colData(cds) row names. Cell names stored in other
cds locations are not modified so you will need to modify them manually
for consistency. Default is FALSE.}

\item{sample_col_name}{A string to be the column name for the colData column
that indicates which original cds the cell derives from. Default is
"sample".}

\item{keep_reduced_dims}{Logical indicating whether to keep the reduced
dimension matrices. Do not keep the reduced dimensions unless you know
that the reduced dimensions are the same in each CDS. This is true for
projected data sets, for example. Default is FALSE.}
}
\value{
A combined cell_data_set object.
}
\description{
This function will combine a list of cell_data_set objects into a new
cell_data_set object.
}
\details{
If any of the counts matrices is BPCells class, the combined
counts matrix will be BPCells class.
}
