% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_models.R
\name{compare_models}
\alias{compare_models}
\title{Compares goodness of fit for two ways of fitting a set of genes' expression}
\usage{
compare_models(model_tbl_full, model_tbl_reduced)
}
\description{
Compares goodness of fit for two ways of fitting a set of genes' expression
}
