% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_models.R
\name{compare_models}
\alias{compare_models}
\title{Compare goodness of fit of two models.}
\usage{
compare_models(model_tbl_full, model_tbl_reduced)
}
\arguments{
\item{model_tbl_full}{A tibble of model objects, generally output of
\code{\link{fit_models}}, to be compared with \code{model_tbl_reduced}}

\item{model_tbl_reduced}{A tibble of model objects, generally output of
\code{\link{fit_models}}, to be compared with \code{model_tbl_full}.}
}
\value{
The result of a likelihood test by gene.
}
\description{
Compares goodness of fit for two ways of fitting a set of genes'
expression using a likelihood ratio test. The likelihood ratio
test helps one decide whether the improvement in fit is large enough
to justify the additional complexity of extra terms in the full
model in comparison to the reduced model.
}
