% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{counts_row_order}
\alias{counts_row_order}
\title{Generic to access cds row order BPCells count matrix.}
\usage{
counts_row_order(x)
}
\arguments{
\item{x}{A cell_data_set object.}
}
\value{
BPCells row order ount matrix.
}
\description{
Generic to access cds row order BPCells count matrix.
}
\examples{
 \donttest{
   cds <- load_a549()
   exprs(cds)
 }         
         
}
