% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detect_genes}
\alias{detect_genes}
\title{Detects genes above minimum threshold.}
\usage{
detect_genes(cds, min_expr = NULL)
}
\arguments{
\item{cds}{the cell_data_set upon which to perform this operation}

\item{min_expr}{the expression threshold}
}
\value{
an updated cell_data_set object
}
\description{
Sets the global expression detection threshold to be used with this cell_data_set
Counts how many cells each feature in a cell_data_set object that are detectably expressed
above a minimum threshold. Also counts the number of genes above this threshold are
detectable in each cell.
}
\examples{
\dontrun{
HSMM <- detect_genes(HSMM, min_expr=0.1)
}
}
