% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimate_size_factors}
\alias{estimate_size_factors}
\title{Function to calculate size factors for single-cell RNA-seq data}
\usage{
estimate_size_factors(
  cds,
  round_exprs = TRUE,
  method = c("mean-geometric-mean-total", "mean-geometric-mean-log-total")
)
}
\arguments{
\item{cds}{The cell_data_set}

\item{round_exprs}{A logic flag to determine whether or not the expression
value should be rounded}

\item{method}{A string to specify the size factor calculation approach.
Options are "mean-geometric-mean-total" (default),
"mean-geometric-mean-log-total".}
}
\value{
Updated cell_data_set object with a new colData column called
'Size_Factor'.
}
\description{
Function to calculate size factors for single-cell RNA-seq data
}
