% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_models.R
\name{evaluate_fits}
\alias{evaluate_fits}
\title{Evaluate the fits of model objects.}
\usage{
evaluate_fits(model_tbl)
}
\arguments{
\item{model_tbl}{A tibble of model objects, generally output of
\code{\link{fit_models}}.}
}
\description{
Evaluate the fits of model objects.
}
\examples{
  cds <- load_worm_embryo()
  cds <- preprocess_cds(cds, num_dim=50)
  cds <- align_cds(cds, alignment_group = "batch", residual_model_formula_str = "~ bg.300.loading + bg.400.loading + bg.500.1.loading + bg.500.2.loading + bg.r17.loading + bg.b01.loading + bg.b02.loading")
  cds <- reduce_dimension(cds)
  ciliated_genes <- c("che-1", "hlh-17", "nhr-6", "dmd-6", "ceh-36", "ham-1")
  cds_subset <- cds[rowData(cds)$gene_short_name \%in\% ciliated_genes,]
  gene_fits <- fit_models(cds_subset, model_formula_str = "~embryo.time")
  evaluate_fits(gene_fits)

}
