% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{fData,cell_data_set-method}
\alias{fData,cell_data_set-method}
\title{Method to access cds rowData table}
\usage{
\S4method{fData}{cell_data_set}(x)
}
\arguments{
\item{x}{A cell_data_set object.}
}
\description{
Method to access cds rowData table
}
\examples{
  \donttest {
    cds <- load_a549()
    fData(cds)
  }

}
