% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{fData<-}
\alias{fData<-}
\title{Generic to set cds rowData table}
\usage{
fData(x) <- value
}
\arguments{
\item{x}{A cell_data_set object.}

\item{value}{A data frame to set to colData table.}
}
\value{
x.
}
\description{
Generic to set cds rowData table
}
\examples{
  \donttest{
    cds <- load_a549()
    fData(cds)[['row_index']] <- seq(nrow(fData(cds)))
  }

}
