% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{fData}
\alias{fData}
\title{Generic to access cds rowData table}
\usage{
fData(x)
}
\arguments{
\item{x}{A cell_data_set object.}
}
\description{
Generic to access cds rowData table
}
\examples{
  cds <- load_a549()
  fData(cds)

}
