% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_genes.R
\name{find_gene_modules}
\alias{find_gene_modules}
\title{Cluster genes into modules that are co-expressed across cells.}
\usage{
find_gene_modules(
  cds,
  reduction_method = c("UMAP"),
  max_components = 2,
  umap.metric = "cosine",
  umap.min_dist = 0.1,
  umap.n_neighbors = 15L,
  umap.fast_sgd = FALSE,
  umap.nn_method = "annoy",
  k = 20,
  leiden_iter = 1,
  partition_qval = 0.05,
  weight = FALSE,
  resolution = NULL,
  random_seed = 0L,
  cores = 1,
  verbose = F,
  preprocess_method = c("PCA", "LSI"),
  nn_control = list(),
  ...
)
}
\arguments{
\item{cds}{the cell_data_set upon which to perform this operation}

\item{reduction_method}{The dimensionality reduction method used to generate the lower dimensional space in which genes will be clustered. Currently only UMAP is supported.}

\item{max_components}{The number of dimensions in which to cluster genes into modules.}

\item{umap.metric}{Metric used by UMAP for measuring similarity between genes .}

\item{umap.min_dist}{Minimum distance parameter passed to UMAP.}

\item{umap.n_neighbors}{Number of nearest neighbors used by UMAP.}

\item{umap.fast_sgd}{Whether to allow UMAP to perform fast stochastic gradient descent. Defaults to TRUE. Setting FALSE will result in slower, but deterministic behavior (if cores=1).}

\item{umap.nn_method}{The method used for nearest neighbor network construction during UMAP.}

\item{k}{number of kNN used in creating the k nearest neighbor graph for Louvain clustering. The number of kNN is related to the resolution of the clustering result, bigger number of kNN gives low resolution and vice versa. Default to be 20}

\item{leiden_iter}{Integer number of iterations used for Leiden clustering. The clustering result with the largest modularity score is used as the final clustering result.  Default to be 1.}

\item{partition_qval}{Significance threshold used in Louvain community graph partitioning.}

\item{weight}{A logic argument to determine whether or not we will use
Jaccard coefficient for two nearest neighbors (based on the overlapping of
their kNN) as the weight used for Louvain clustering. Default to be FALSE.}

\item{resolution}{Resolution parameter passed to Louvain. Can be a list. If
so, this method will evaluate modularity at each resolution and use the
one with the highest value.}

\item{random_seed}{the seed used by the random number generator in Leiden.}

\item{cores}{number of cores computer should use to execute function}

\item{verbose}{Whether or not verbose output is printed.}

\item{preprocess_method}{a string specifying the low-dimensional space
to use for gene loadings, currently either PCA or LSI. Default is
"PCA".}

\item{nn_control}{list See set_nn_control for a description of available
and default list values.}

\item{...}{Additional arguments passed to UMAP and Louvain analysis.}
}
\value{
A dataframe with genes and the modules to which they are assigned.
}
\description{
Cluster genes into modules that are co-expressed across cells.
}
