% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_models.R
\name{fit_model_helper}
\alias{fit_model_helper}
\title{Helper function for model fitting}
\usage{
fit_model_helper(x, model_formula_str, expression_family,
  disp_func = NULL, clean_model = TRUE, verbose = FALSE, ...)
}
\arguments{
\item{x}{test}

\item{model_formula_str}{a formula string specifying the model to fit for the genes.}

\item{expression_family}{specifies the family function used for expression responses}

\item{disp_func}{test}

\item{verbose}{Whether to show VGAM errors and warnings. Only valid for cores = 1.}

\item{...}{test}
}
\description{
test
}
\keyword{internal}
