% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_models.R
\name{fit_models}
\alias{fit_models}
\title{Fits a model for each gene in a cell_data_set object.}
\usage{
fit_models(cds, model_formula_str, expression_family = "quasipoisson",
  cores = 1, clean_model = TRUE, verbose = FALSE, ...)
}
\arguments{
\item{cds}{the cell_data_set upon which to perform this operation}

\item{model_formula_str}{a formula string specifying the model to fit for the genes.}

\item{expression_family}{specifies the family function used for expression responses.}

\item{cores}{the number of processor cores to be used during fitting.}
}
\value{
a tibble containing model objects
}
\description{
This function fits a generalized linear model for each gene in a cell_data_set.
Formulae can be provided to account for
additional covariates (e.g. day collected, genotype of cells, media conditions, etc).
}
