% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_transfer.R
\name{fix_missing_cell_labels}
\alias{fix_missing_cell_labels}
\title{Try to replace NA cell column data values.}
\usage{
fix_missing_cell_labels(
  cds,
  reduction_method = c("UMAP", "PCA", "LSI"),
  column_name,
  out_notna_models_dir = NULL,
  k = 10,
  nn_control = list(),
  top_frac_threshold = 0.5,
  top_next_ratio_threshold = 1.5,
  verbose = FALSE
)
}
\arguments{
\item{reduction_method}{a string specifying the reduced dimension
matrix to use for the label transfer. These are "PCA",
"LSI", and "UMAP". Default is "UMAP".}

\item{column_name}{a string giving the name of the query
cds column to fix.}

\item{out_notna_models_dir}{a string giving the name of the transform
model directory to which to save the not-NA transform models,
which includes the nearest neighbor index.}

\item{k}{an integer giving the number of reference nearest
neighbors to find. This value must be large enough to find
meaningful column value fractions. The default is 10.}

\item{nn_control}{a list of parameters used to make the nearest
neighbors index. See the set_nn_control help for additional
details. The default is to use the global nn_control list.}

\item{top_frac_threshold}{a numeric value giving the minimum
value of the top fraction of reference values required
for transfering the reference value to the query. The top
fraction is the fraction of the k neighbors with the most
frequent value. The default is 0.5.}

\item{top_next_ratio_threshold}{a numeric value giving the
minimum value ratio of the counts of the most frequent to the
second most frequent reference values. The default is 1.5.}

\item{verbose}{a boolean controlling verbose output.}

\item{cds_query}{the cell_data_set upon which to perform this operation}
}
\value{
an updated cell_data_set object
}
\description{
Try to replace NA values left in a query cds by
transfer_cell_labels().

fix_missing_cell_labels() uses non-NA cell data values in the query cds
to replace NAs in nearby cells. It partitions the cells into a set with
NA and a set with non-NA column data values. It makes a nearest
neighbor index using cells with non-NA values, and for each cell
with NA, it tries to find an acceptable non-NA column data value as
follows. If at least top_frac_threshold fraction of them have the same
value, it replaces the NA with it. If not, it checks whether the ratio
of the most frequent to the second most frequent values is at least
top_next_ratio_threshold, in which case it copies the most frequent
value. Otherwise, it leaves the NA.
}
