% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_markers.R
\name{generate_garnett_marker_file}
\alias{generate_garnett_marker_file}
\title{Title}
\usage{
generate_garnett_marker_file(marker_test_res, file = "./marker_file.txt",
  max_genes_per_group = 10)
}
\arguments{
\item{marker_test_res}{Tibble of top markers, output of
\code{\link{top_markers}}.}

\item{file}{Path to the marker file to be generated. Default is
"./marker_file.txt".}

\item{max_genes_per_group}{Numeric, the maximum number of genes to output
per cell type entry. Default is 10.}
}
\value{
None, marker file is written to \code{file} parameter location.
}
\description{
Title
}
