% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_markers.R
\name{generate_garnett_marker_file}
\alias{generate_garnett_marker_file}
\title{Generate a Garnett marker file from top_markers output.}
\usage{
generate_garnett_marker_file(top_markers, file = "./marker_file.txt",
  qval_cutoff = 0.05, max_genes_per_group = 10)
}
\arguments{
\item{top_markers}{Tibble of top markers, output of
\code{\link{top_markers}}.}

\item{file}{Path to the marker file to be generated. Default is
"./marker_file.txt".}

\item{qval_cutoff}{Numeric, the q-value cutoff below which to include genes
as markers. Default is 0.05.}

\item{max_genes_per_group}{Numeric, the maximum number of genes to output
per cell type entry. Default is 10.}
}
\value{
None, marker file is written to \code{file} parameter location.
}
\description{
Generate a Garnett marker file from top_markers output.
}
