% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identity.R
\name{identity_table}
\alias{identity_table}
\title{Report the matrix and model identity information.}
\usage{
identity_table(cds)
}
\arguments{
\item{cds}{the cell_data_set to use.}
}
\value{
Write identity information to stdout.
}
\description{
Write the cell_data_set matrix and model identity
information to stdout.
}
\details{
A matrix identity is a time stamp that is made when
a reduced dimension matrix is created. The time stamp is an
MD5 checksum string made from the time and a counter, which
make it unique. Some functions that read count
matrices also make a matrix identity, for example,
load_mm_data(). In addition to the matrix identity for the
reduced dimension matrix, Monocle3 stores the identity of the
matrix from which the matrix was made. For example, in the
case of a UMAP reduced dimension matrix made from a PCA reduced
dimension matrix, the cds has the identity of both the UMAP
and the PCA matrices. The identity_table() function stores
the UMAP identity as 'matrix_id' and the PCA identity as
'prev_matrix_id'. This gives a way to trace a matrix to its
origin. The ability to trace the matrix may be helpful when
a cds is saved and restored, and then partially reprocessed.

The matrix identity string is stored in the internal metadata
slot of the cell_data_set.

Note that the same matrix created at different times will
have different matrix identities. We chose to use a time
stamp because the alternative of making a checksum from a
very large matrix can take a long time.
}
