% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identity.R
\name{identity_table}
\alias{identity_table}
\title{Report matrix and model identity information.}
\usage{
identity_table(cds)
}
\arguments{
\item{cds}{the cell_data_set to use.}
}
\value{
Write identity information to stdout.
}
\description{
Write the cell_data_set matrix and model
identity information to stdout.
}
\details{
A matrix identity is a checksum that is stored
in the cell_data_set when a reduced dimension matrix is
created and when certain functions read count matrices
into the cell_data_set, such as load_mm_data(). At the
same time, the same checksum is stored as the model
identity in order to link the model to its matrix.

Additionally, Monocle3 stores the identity of the matrix
from which the matrix was made. For example, in the case
of a UMAP reduced dimension matrix made from a PCA
reduced dimension matrix, the cell_data_set has the
identities of both the UMAP and the PCA matrices. The
UMAP identity is stored as 'matrix_id' and the PCA as
'prev_matrix_id'. Similarly, the model and the previous
model identities are stored as 'model_id' and
'prev_model_id'. This allows one to trace a matrix
to its origin, which may be helpful when a cell_data_set
is partially reprocessed; for example, if preprocess_cds()
is re-run but reduce_dimension() is not. Also, it may be
helpful when transform models are loaded with the
load_transform_models() function, in which case
the matrix and model identities will differ.

The identity of the model used to transform a matrix
is stored with the matrix identity information as
'model_id'. Ordinarily, the matrix 'matrix_id' and
'model_id' and the corresponding model 'model_id' will
have the same string value. However, they differ when
the preprocess_transform() and reduce_dim_transform()
functions are used to transform a matrix.

Notes:
\itemize{
\item Certain file and directory paths may be stored in the
cell_data_set as identifiers.
\item Checksums are calculated using the digest function in
the digest package. The matrix dimensions are stored
with the checksum.
\item Matrix transformations such as subsetting and row and
or column reordering do not affect the matrix identity.
\item The matrix identity string is stored in the internal
metadata slot of the cell_data_set and the model
identity string is stored in the model object in the
cds@reduce_dim_aux slot of the cell_data_set.
}
}
