% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_mm_data}
\alias{load_mm_data}
\title{Load data from matrix market format files.}
\usage{
load_mm_data(
  mat_path,
  feature_anno_path,
  cell_anno_path,
  header = FALSE,
  feature_metadata_column_names = NULL,
  cell_metadata_column_names = NULL,
  umi_cutoff = 100,
  sep = "\\t"
)
}
\arguments{
\item{mat_path}{Path to the Matrix Market .mtx matrix file. The
values are read and stored  as a sparse matrix with nrows and ncols,
as inferred from the file.}

\item{feature_anno_path}{Path to a feature annotation file. The
feature_anno_path file must have nrows lines and at least one column.
The values in the first column label the matrix rows and each must be
distinct in the column. Values in additional columns are stored in
the cell_data_set 'gene' metadata. For gene features, we urge use of
official gene IDs for labels, such as Ensembl or Wormbase IDs. In this
case, the second column has typically a 'short' gene name in the second
column. Additional information such as gene_biotype may be stored in
additional columns starting with column 3. Required.}

\item{cell_anno_path}{Path to a cell annotation file. The cell_anno_path
file must have ncols lines and at least one column. The values in the
first column label the matrix columns and each must be distinct in the
column. Values in additional columns are stored in the cell_data_set
cells metadata. Required.}

\item{header}{Logical set to TRUE if both feature_anno_path and
cell_anno_path files have column headers, or set to FALSE if both
files do not have column headers (only these cases are supported).
The files may have either ncols or ncols-1 header fields. In both
cases, the first column is used as the matrix dimension names. The
default is FALSE.}

\item{feature_metadata_column_names}{A character vector of feature
metadata column names. The number of names must one less than the
number of columns in the feature_anno_path file. For no feature
metadata column names, set feature_metadata_column_names to NULL.
These values replace the feature_anno_path file header values. The
default is NULL.}

\item{cell_metadata_column_names}{A character vector of cell
metadata column names. The number of names must one less than the
number of columns in the cell_anno_path file. For no cell metadata
column names, set cell_metadata_column_names to NULL. These values
replace the cell_anno_path file header values. The default is NULL.}

\item{umi_cutoff}{UMI per cell cutoff. Columns (cells) with less
than umi_cutoff total counts are removed from the matrix. The
default is 100.}

\item{sep}{field separator character in annotation files. The
default is the tab character for tab-separated-value files.

Notes:
o  this function estimates size factors.}
}
\value{
cds object
}
\description{
Load data from matrix market format files.
}
