% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_worm_embryo}
\alias{load_worm_embryo}
\title{Build a cell_data_set from the data stored in inst/extdata directory.}
\usage{
load_worm_embryo()
}
\description{
Build a cell_data_set from the data stored in inst/extdata directory.
}
\keyword{internal}
