% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_models.R
\name{model_predictions}
\alias{model_predictions}
\title{Predict output of fitted models and return as a matrix}
\usage{
model_predictions(model_tbl, new_data, type = "response")
}
\description{
Predict output of fitted models and return as a matrix
}
