% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_data_set.R
\name{new_cell_data_set}
\alias{new_cell_data_set}
\title{Create a new cell_data_set object.}
\usage{
new_cell_data_set(expression_data, cell_metadata = NULL, gene_metadata = NULL)
}
\arguments{
\item{expression_data}{expression data matrix for an experiment, can be a
sparseMatrix.}

\item{cell_metadata}{data frame containing attributes of individual cells,
where \code{row.names(cell_metadata) = colnames(expression_data)}.}

\item{gene_metadata}{data frame containing attributes of features
(e.g. genes), where
\code{row.names(gene_metadata) = row.names(expression_data)}.}
}
\value{
a new cell_data_set object
}
\description{
Create a new cell_data_set object.
}
\examples{
small_a549_colData_df <- readRDS(system.file("extdata",
                                             "small_a549_dex_pdata.rda",
                                             package = "monocle3"))
small_a549_rowData_df <- readRDS(system.file("extdata",
                                             "small_a549_dex_fdata.rda",
                                             package = "monocle3"))
small_a549_exprs <- readRDS(system.file("extdata",
                                        "small_a549_dex_exprs.rda",
                                        package = "monocle3"))
small_a549_exprs <- small_a549_exprs[,row.names(small_a549_colData_df)]

cds <- new_cell_data_set(expression_data = small_a549_exprs,
                         cell_metadata = small_a549_colData_df,
                         gene_metadata = small_a549_rowData_df)

}
