% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_cells.R
\name{order_cells}
\alias{order_cells}
\title{Orders cells according to pseudotime.}
\usage{
order_cells(cds, root_pr_nodes = NULL, reduction_method = "UMAP",
  root_cells = NULL, reverse = FALSE, orthogonal_proj_tip = FALSE,
  verbose = FALSE)
}
\arguments{
\item{cds}{the cell_data_set upon which to perform this operation}

\item{root_pr_nodes}{The starting principal points. We learn a principal graph that passes through the middle of the data points and use it to represent the developmental process.}

\item{root_cells}{The starting cells. Each cell corresponds to a principal point and multiple cells can correspond to the same principal point.}

\item{reverse}{Whether to reverse the direction of the trajectory}

\item{orthogonal_proj_tip}{Whether to perform orthogonal projection for cells corresponding to the tip principal points. Default to be FALSE}

\item{verbose}{Whether to show running information for order_cells}
}
\value{
an updated cell_data_set object, in which phenoData contains values for State and Pseudotime for each cell
}
\description{
Learns a "trajectory" describing the biological process the cells are
going through, and calculates where each cell falls within that trajectory.
Monocle learns trajectories in two steps. The first step is reducing the dimensionality
of the data with \code{\link{reduce_dimension}()}. The second is this function.
function. This function takes as input a cell_data_set and returns it with
two new columns: \code{Pseudotime} and \code{State}, which together encode
where each cell maps to the trajectory. \code{order_cells()} optionally takes
a "root" state, which you can use to specify the start of the trajectory. If
you don't provide a root state, one is selected arbitrarily.
}
\details{
The \code{reduction_method} argument to \code{\link{reduce_dimension}()}
determines which algorithm is used by \code{order_cells()} to learn the trajectory.
If \code{reduction_method == "ICA"}, this function uses \emph{polygonal reconstruction}
to learn the underlying trajectory. If \code{reduction_method == "DDRTree"},
the trajectory is specified by the principal graph learned by the
\code{\link[DDRTree]{DDRTree}()} function.

Whichever algorithm you use, the trajectory will be composed of segments.
The cells from a segment will share the same value of \code{State}. One of
these segments will be selected as the root of the trajectory arbitrarily.
The most distal cell on that segment will be chosen as the "first" cell in the
trajectory, and will have a Pseudotime value of zero. \code{order_cells()} will
then "walk" along the trajectory, and as it encounters additional cells, it
will assign them increasingly large values of Pseudotime.
}
