% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_cells.R
\name{order_cells}
\alias{order_cells}
\title{Orders cells according to pseudotime.}
\usage{
order_cells(
  cds,
  reduction_method = "UMAP",
  root_pr_nodes = NULL,
  root_cells = NULL,
  verbose = FALSE
)
}
\arguments{
\item{cds}{the cell_data_set upon which to perform this operation}

\item{reduction_method}{a string specifying the reduced dimension method to
use when ordering cells. Currently only "UMAP" is supported.}

\item{root_pr_nodes}{NULL or a vector of starting principal points. If
provided, pseudotime will start (i.e. be zero) at these graph nodes. You
can find the principal point names by running plot_cells with
label_principal_points = TRUE. Both \code{root_pr_nodes} and
\code{root_cells} cannot be provided.}

\item{root_cells}{NULL or a vector of starting cells. If provided,
pseudotime will start (i.e. be zero) at these cells. Both
\code{root_pr_nodes} and \code{root_cells} cannot be provided.}

\item{verbose}{Whether to show running information for order_cells}
}
\value{
an updated cell_data_set object.
}
\description{
Assigns cells a pseudotime value based on their projection on the principal
graph learned in the \code{learn_graph} function and the position of chosen
root states. This function takes as input a cell_data_set and returns it
with pseudotime information stored internally.
\code{order_cells()} optionally takes "root" state(s) in the form of cell
or principal graph node IDs, which you can use to specify the start of the
trajectory. If you don't provide a root state, an plot will be generated
where you can choose the root state(s) interactively. The trajectory will be
composed of segments.
}
\examples{
  \donttest {
    cds <- load_worm_embryo()
    cds <- preprocess_cds(cds, num_dim=50)
    cds <- align_cds(cds, alignment_group = "batch", residual_model_formula_str = "~ bg.300.loading + bg.400.loading + bg.500.1.loading + bg.500.2.loading + bg.r17.loading + bg.b01.loading + bg.b02.loading")
    cds <- reduce_dimension(cds)
    cds <- cluster_cells(cds)
    cds <- learn_graph(cds)
    cds <- order_cells(cds)
  }

}
