% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{partitions,cell_data_set-method}
\alias{partitions,cell_data_set-method}
\title{Method to extract partitions from CDS object}
\usage{
\S4method{partitions}{cell_data_set}(x, reduction_method = "UMAP")
}
\arguments{
\item{x}{A cell_data_set object.}

\item{reduction_method}{Reduced dimension to partitions clusters for.}
}
\description{
Method to extract partitions from CDS object
}
