% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{partitions,cell_data_set-method}
\alias{partitions,cell_data_set-method}
\title{Method to extract partitions from CDS object}
\usage{
\S4method{partitions}{cell_data_set}(x, reduction_method = "UMAP")
}
\arguments{
\item{x}{A cell_data_set object.}

\item{reduction_method}{Reduced dimension to partitions clusters for.}
}
\description{
Method to extract partitions from CDS object
}
\examples{
  \donttest{
    cds <- load_worm_embryo()
    cds <- preprocess_cds(cds)
    cds <- reduce_dimension(cds)
    cds <- cluster_cells(cds)
    partitions_factors <- partitions(cds, "UMAP")
  }

}
