% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_cells}
\alias{plot_cells}
\title{Plots the cells along with their trajectories.}
\usage{
plot_cells(cds, x = 1, y = 2, reduction_method = "UMAP",
  color_by = "Cluster", genes = NULL, show_trajectory_graph = TRUE,
  trajectory_graph_color = "black",
  trajectory_graph_segment_size = 0.75, norm_method = c("log",
  "size_only"), label_cell_groups = TRUE,
  label_groups_by_cluster = TRUE, group_label_size = 2,
  labels_per_group = 1, label_branch_points = TRUE,
  label_roots = TRUE, label_leaves = TRUE, graph_label_size = 2,
  cell_size = 0.35, alpha = 1, min_expr = 0.1, rasterize = TRUE)
}
\arguments{
\item{cds}{cell_data_set for the experiment}

\item{x}{the column of reducedDimS(cds) to plot on the horizontal axis}

\item{y}{the column of reducedDimS(cds) to plot on the vertical axis}

\item{color_by}{the cell attribute (e.g. the column of colData(cds)) to map to each cell's color}

\item{cell_size}{The size of the point for each cell}

\item{alpha}{The alpha aesthetics for the original cell points, useful to highlight the learned principal graph}

\item{show_backbone}{whether to show the diameter path of the MST used to order the cells}

\item{backbone_color}{the color used to render the backbone.}

\item{markers}{a gene name or gene id to use for setting the size of each cell in the plot}

\item{markers_linear}{a boolean used to indicate whether you want to scale the markers logarithimically or linearly}

\item{show_cell_names}{draw the name of each cell in the plot}

\item{show_state_number}{show state number}

\item{cell_link_size}{The size of the line segments connecting cells (when used with ICA) or the principal graph (when used with DDRTree)}

\item{cell_name_size}{the size of cell name labels}

\item{state_number_size}{the size of the state number}

\item{show_branch_points}{Whether to show icons for each branch point (only available after running assign_cell_states)}

\item{theta}{How many degrees you want to rotate the trajectory}

\item{...}{Additional arguments passed into scale_color_viridis function}
}
\value{
a ggplot2 plot object
}
\description{
Plots the cells along with their trajectories.
}
\examples{
\dontrun{
lung <- load_lung()
plot_cell_trajectory(lung)
plot_cell_trajectory(lung, color_by="Pseudotime", show_backbone=FALSE)
plot_cell_trajectory(lung, markers="MYH3")
}
}
