% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_percent_cells_positive}
\alias{plot_percent_cells_positive}
\title{Plots the number of cells expressing one or more genes above a given value
as a barplot}
\usage{
plot_percent_cells_positive(cds_subset, group_cells_by = NULL,
  min_expr = 0, nrow = NULL, ncol = 1, panel_order = NULL,
  plot_as_count = FALSE, label_by_short_name = TRUE,
  normalize = TRUE, plot_limits = NULL)
}
\arguments{
\item{cds_subset}{Subset cell_data_set to be plotted.}

\item{group_cells_by}{the cell attribute (e.g. the column of colData(cds))
to group cells by on the horizontal axis. If NULL, all cells plotted as
one group.}

\item{min_expr}{the minimum (untransformed) expression level to consider the
gene 'expressed'. Default is 0.}

\item{nrow}{the number of panels per row in the figure.}

\item{ncol}{the number of panels per column in the figure.}

\item{panel_order}{the order in which genes should be laid out
(left-to-right, top-to-bottom). Should be gene_short_name if
\code{label_by_short_name = TRUE} or feature ID if
\code{label_by_short_name = FALSE}.}

\item{plot_as_count}{Logical, whether to plot as a count of cells rather
than a percent. Default is FALSE.}

\item{label_by_short_name}{label figure panels by gene_short_name (TRUE) or
feature id (FALSE). Default is TRUE.}

\item{normalize}{Logical, whether or not to normalize expression by size
factor. Default is TRUE.}

\item{plot_limits}{A pair of number specifying the limits of the y axis. If
\code{NULL}, scale to the range of the data. Example \code{c(0,100)}.}
}
\value{
a ggplot2 plot object
}
\description{
@description Accepts a subset cell_data_set and the parameter
\code{group_cells_by}, used for dividing cells into groups. Returns one or
more bar graphs (one graph for each gene in the cell_data_set). Each graph
shows the percentage (or number) of cells that express a gene in each
sub-group in the cell_data_set.
}
\examples{
cds <- load_a549()
cds_subset <- cds[row.names(subset(rowData(cds),
                                  gene_short_name \%in\% c("NDRG4", "HBG2"))),]
plot_percent_cells_positive(cds_subset, group_cells_by="culture_plate")
}
