% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_umi_per_cell}
\alias{plot_umi_per_cell}
\title{Plot a histogram of the number of UMIs per cell with a vertical line showing the cutoff for the minimum number of UMIs.}
\usage{
plot_umi_per_cell(cds, min_rna_umi = 100)
}
\arguments{
\item{cds}{A cell_data_set for plotting.}

\item{min_rna_umi}{Cutoff for the minimum number of UMIs per cell.}
}
\value{
A ggplot2 object.
}
\description{
Plot a histogram of the number of UMIs per cell with a vertical line showing the cutoff for the minimum number of UMIs.
}
