% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_cds.R
\name{preprocess_cds}
\alias{preprocess_cds}
\title{Preprocess a cds to prepare for trajectory inference}
\usage{
preprocess_cds(
  cds,
  method = c("PCA", "LSI"),
  num_dim = 50,
  norm_method = c("log", "size_only", "none"),
  use_genes = NULL,
  pseudo_count = NULL,
  scaling = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{cds}{the cell_data_set upon which to perform this operation}

\item{method}{a string specifying the initial dimension method to use,
currently either PCA or LSI. For LSI (latent semantic indexing), it
converts the (sparse) expression matrix into tf-idf matrix and then
performs SVD to decompose the gene expression / cells into certain
modules / topics. Default is "PCA".}

\item{num_dim}{the dimensionality of the reduced space.}

\item{norm_method}{Determines how to transform expression values prior to
reducing dimensionality. Options are "log", "size_only", and "none".
Default is "log". Users should only use "none" if they are confident that
their data is already normalized.}

\item{use_genes}{NULL or a list of gene IDs. If a list of gene IDs, only
this subset of genes is used for dimensionality reduction. Default is
NULL.}

\item{pseudo_count}{NULL or the amount to increase expression values before
normalization and dimensionality reduction. If NULL (default), a
pseudo_count of 1 is added for log normalization and 0 is added for size
factor only normalization.}

\item{scaling}{When this argument is set to TRUE (default), it will scale
each gene before running trajectory reconstruction. Relevant for
method = PCA only.}

\item{verbose}{Whether to emit verbose output during dimensionality
reduction}

\item{...}{additional arguments to pass to limma::lmFit if
residual_model_formula is not NULL}
}
\value{
an updated cell_data_set object
}
\description{
Most analyses (including trajectory inference, and clustering)
in Monocle3, require various normalization and preprocessing steps.
\code{preprocess_cds} executes and stores these preprocessing steps.

Specifically, depending on the options selected, \code{preprocess_cds} first
normalizes the data by log and size factor to address depth differences, or
by size factor only. Next, \code{preprocess_cds} calculates a lower
dimensional space that will be used as the input for further dimensionality
reduction like tSNE and UMAP.
}
