% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{preprocess_transform}
\alias{preprocess_transform}
\title{Apply a preprocess transform model to a cell_data_set.}
\usage{
preprocess_transform(cds, method = c("PCA"), block_size = NULL, cores = 1)
}
\arguments{
\item{cds}{A cell_data_set to be transformed.}

\item{method}{A previously loaded transform model that
is used to reduce the dimensions of the count matrix
in cds.}

\item{block_size}{A numeric value for the DelayedArray
block size used only in this function. Default is
NULL, which does not affect the current block size.}

\item{cores}{The number of cores to use for the matrix multiplication.}
}
\value{
A cell_data_set with a preprocess reduced count
matrix.
}
\description{
Applies a previously calculated preprocess transform
model to a new count matrix. For more information read
the help information for save_transform_models.
}
