% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{preprocess_transform}
\alias{preprocess_transform}
\title{Apply a preprocess transform model to a cell_data_set.}
\usage{
preprocess_transform(cds, method = c("PCA", "LSI"))
}
\arguments{
\item{cds}{A cell_data_set to be transformed.}

\item{method}{A previously loaded transform model that
is used to reduce the dimensions of the count matrix
in cds.}
}
\value{
A cell_data_set with a preprocess reduced count
matrix.
}
\description{
Applies a previously calculated preprocess transform
model to a new count matrix. For more information read
the help information for save_transform_models.
}
