% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{principal_graph<-,cell_data_set-method}
\alias{principal_graph<-,cell_data_set-method}
\title{Generic to set principal graph to CDS}
\usage{
\S4method{principal_graph}{cell_data_set}(x) <- value
}
\arguments{
\item{x}{A cell_data_set object.}

\item{value}{A principal graph object.}
}
\description{
Generic to set principal graph to CDS
}
\examples{
 \donttest{
    cell_metadata <- readRDS(system.file('extdata', 'worm_embryo/worm_embryo_coldata.rds', package='monocle3'))
    gene_metadata <- readRDS(system.file('extdata', 'worm_embryo/worm_embryo_rowdata.rds', package='monocle3'))
    expression_matrix <- readRDS(system.file('extdata', 'worm_embryo/worm_embryo_expression_matrix.rds', package='monocle3'))

    cds <- new_cell_data_set(expression_data=expression_matrix,
                             cell_metadata=cell_metadata,
                             gene_metadata=gene_metadata)

   cds <- preprocess_cds(cds, num_dim=50)
   cds <- align_cds(cds, alignment_group = "batch", residual_model_formula_str = "~ bg.300.loading + bg.400.loading + bg.500.1.loading + bg.500.2.loading + bg.r17.loading + bg.b01.loading + bg.b02.loading")
   cds <- reduce_dimension(cds)
   ciliated_genes <- c("che-1", "hlh-17", "nhr-6", "dmd-6", "ceh-36", "ham-1")
   cds <- cluster_cells(cds)
   cds <- learn_graph(cds)
   pr_gr <- principal_graph(cds)
   principal_graph(cds) <- NULL
   principal_graph(cds) <- pr_gr
 }

}
