% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{principal_graph<-}
\alias{principal_graph<-}
\title{Generic to set principal graph to CDS}
\usage{
principal_graph(x) <- value
}
\arguments{
\item{x}{A cell_data_set object.}

\item{value}{A principal graph object.}
}
\description{
Generic to set principal graph to CDS
}
