% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-cell_data_set.R
\name{principal_graph}
\alias{principal_graph}
\title{Retrieves the minimum spanning tree generated by Monocle during cell ordering.}
\usage{
principal_graph(cds)
}
\arguments{
\item{cds}{expression data matrix for an experiment}
}
\value{
An igraph object representing the cell_data_set's minimum spanning tree.
}
\description{
Retrieves the minimum spanning tree (MST) that Monocle constructs during orderCells().
This MST is mostly used in plot_spanning_tree to help assess the accuracy
of Monocle's ordering.
}
\examples{
\dontrun{
T <- principal_graph(HSMM)
}
}
