% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{principal_graph_aux<-,cell_data_set-method}
\alias{principal_graph_aux<-,cell_data_set-method}
\title{Method to set principal graph auxiliary information into CDS}
\usage{
\S4method{principal_graph_aux}{cell_data_set}(x) <- value
}
\arguments{
\item{x}{A cell_data_set object.}

\item{value}{A SimpleList of principal graph auxiliary information.}
}
\description{
Method to set principal graph auxiliary information into CDS
}
\examples{
 \donttest{
   cds <- load_worm_embryo()
   cds <- preprocess_cds(cds, num_dim=50)
   cds <- align_cds(cds, alignment_group = "batch", residual_model_formula_str = "~ bg.300.loading + bg.400.loading + bg.500.1.loading + bg.500.2.loading + bg.r17.loading + bg.b01.loading + bg.b02.loading")
   cds <- reduce_dimension(cds)
   ciliated_genes <- c("che-1", "hlh-17", "nhr-6", "dmd-6", "ceh-36", "ham-1")
   cds <- cluster_cells(cds)
   cds <- learn_graph(cds)
   pr_gr_aux <- principal_graph_aux(cds)
   principal_graph_aux(cds) <- NULL
   principal_graph_aux(cds) <- pr_gr_aux
 }

}
