% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{principal_graph_aux}
\alias{principal_graph_aux}
\title{Generic to extract principal graph auxiliary information from CDS}
\usage{
principal_graph_aux(x)
}
\arguments{
\item{x}{A cell_data_set object.}
}
\description{
Generic to extract principal graph auxiliary information from CDS
}
