% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{pseudotime,cell_data_set-method}
\alias{pseudotime,cell_data_set-method}
\title{Method to extract pseudotime from CDS object}
\usage{
\S4method{pseudotime}{cell_data_set}(x, reduction_method = "UMAP")
}
\arguments{
\item{x}{A cell_data_set object.}

\item{reduction_method}{Reduced dimension to extract clusters for.}
}
\description{
Method to extract pseudotime from CDS object
}
\examples{
 \donttest{
   cds <- load_worm_embryo()
   cds <- preprocess_cds(cds, num_dim=50)
   cds <- align_cds(cds, alignment_group = "batch", residual_model_formula_str = "~ bg.300.loading + bg.400.loading + bg.500.1.loading + bg.500.2.loading + bg.r17.loading + bg.b01.loading + bg.b02.loading")
   cds <- reduce_dimension(cds)
   ciliated_genes <- c("che-1", "hlh-17", "nhr-6", "dmd-6", "ceh-36", "ham-1")
   cds <- cluster_cells(cds)
   cds <- learn_graph(cds)
   cds <- order_cells(cds)
   ps_tim <- pseudotime(cds)
 }

}
