% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reduce_dimension_transform}
\alias{reduce_dimension_transform}
\title{Apply a reduce_transform transform model to a cell_data_set.}
\usage{
reduce_dimension_transform(cds, method = c("UMAP"))
}
\arguments{
\item{cds}{A cell_data_set to be transformed.}

\item{method}{A previously loaded transform model that
is used to reduce the dimensions of the preprocessed
count matrix in cds.}
}
\value{
A cell_data_set with a reduce_dimension transformed
reduced count matrix.
}
\description{
Applies a previously calculated reduce_dimension transform
model to a new preprocess transformed matrix. For more
information read the help information for
save_transform_models.
}
