% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{save_transform_models}
\alias{save_transform_models}
\title{Save the set of cell_data_set transform models.}
\usage{
save_transform_models(cds, directory_path, comment = "", verbose = TRUE)
}
\arguments{
\item{cds}{A cell_data_set with existing models.}

\item{directory_path}{A string giving the name of the directory
in which to write the model files.}

\item{comment}{A string with optional notes that is saved with
the objects.}

\item{verbose}{A boolean determining whether to print information
about the saved files.}
}
\value{

}
\description{
Save the transform models to a specified directory
by writing the R objects to RDS files and
the Annoy nearest neighbor indexes to individual index files.
save_transform_models saves transforms made by the
preprocess_cds, align_cds, and reduce_dimension functions.
From the preprocess_cds, align_cds, and reduce_dimension
transforms, save_transform_models saves the objects required
to transform new count matrices into their reduced dimension
spaces. If the preprocess_cds, align_cds, or reduce_dimension
functions are run with the build_nn_index=TRUE parameter,
those nearest neighbor indexes are saved as well. The model
transforms are made by running preprocess_cds, align_cds,
and reduce_dimension on a primary data set and saving the
the transform models. In order to transform a new count
matrix, load it into a new cell_data_set, load the saved
transform models, and apply the models to the new count
matrix without running preprocess_cds, align_cds, or
reduce_dimension.
}
