% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{save_transform_models}
\alias{save_transform_models}
\title{Save cell_data_set transform models.}
\usage{
save_transform_models(
  cds,
  directory_path,
  comment = "",
  verbose = TRUE,
  archive_control = list(archive_type = "tar", archive_compression = "none")
)
}
\arguments{
\item{cds}{a cell_data_set with existing models.}

\item{directory_path}{a string giving the name of the directory
in which to write the model files.}

\item{comment}{a string with optional notes that is saved with
the objects.}

\item{verbose}{a boolean determining whether to print information
about the saved files.}
}
\value{
none.
}
\description{
Save the transform models in the cell_data_set to the
specified directory by writing the R objects to RDS
files and the nearest neighbor indexes to
index files. save_transform_models saves transform
models made by running the preprocess_cds and
reduce_dimension functions on an initial cell_data_set.
Subsequent cell_data_sets are transformed into the
reduced dimension space of the initial cell_data_set by
loading the new data into a new cell_data_set, loading
the initial data set transform models into the new
cell_data_set using the load_transform_models function,
and applying those transform models to the new data set
using the preprocess_transform and
reduce_dimension_transform functions. In this case, do
not run the preprocess_cds or reduce_dimension
functions on the new cell_data_set. Additionally,
save_transform_models saves nearest neighbor indexes
when the preprocess_cds and reduce_dimension
functions are run with the make_nn_index=TRUE parameter.
These indexes are used to find matches between cells in
the new processed cell_data_set and the initial
cell_data_set using index search functions. For more
information see the help for transfer_cell_labels.
save_transform_models saves the models to a directory
given by directory_path.
}
\examples{
  \dontrun{
    cds <- load_a549()
    cds <- preprocess_cds(cds)
    cds <- reduce_dimension(cds)
    save_transform_models(cds, 'tm')
  }

}
