% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{search_nn_index}
\alias{search_nn_index}
\title{Search Annoy index for cells near the cells in cds.}
\usage{
search_nn_index(
  cds,
  stage = c("PCA", "align", "reduce_dimension"),
  method = c("PCA", "LSI", "Aligned", "UMAP"),
  n = 5,
  search_k = 100 * n,
  ...
)
}
\description{
Search the Annoy nearest neighbor index for cells near cells in cds.
}
\details{
param cds A cell_data_set.
param stage The stage in which the Annoy index was made.
param method The method for which the Annoy index was made. This
can be method 'PCA' or 'LSI' for stage 'preprocess', method 'Aligned'
for stage 'Align', and method 'UMAP' for stage 'reduce_dimension'.
param n An integer for the number of nearest neighbors to return for
each cell.
param search_k An integer used to balance accuracy and speed. Larger
values give more accurate results.
param ... Parameters to pass through to annoy_search.

export
}
