% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_neighbors.R
\name{search_nn_index}
\alias{search_nn_index}
\title{Search index for cells near the cells in cds.}
\usage{
search_nn_index(
  cds,
  reduction_method = c("PCA", "LSI", "Aligned", "tSNE", "UMAP"),
  k,
  nn_control
)
}
\arguments{
\item{cds}{A cell_data_set.}

\item{reduction_method}{The method for which the index was made. method
can be 'PCA', 'LSI','Aligned', 'tSNE', or 'UMAP'.}

\item{k}{An integer for the number of nearest neighbors to return for
each cell. Default is 5.}

\item{...}{Parameters to pass through to annoy_search.}
}
\description{
Search the nearest neighbor index for cells near cells in cds.
}
