% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_neighbors.R
\name{search_nn_index}
\alias{search_nn_index}
\title{Search Annoy index for cells near the cells in cds.}
\usage{
search_nn_index(
  cds,
  method = c("PCA", "LSI", "Aligned", "tSNE", "UMAP"),
  n = 5,
  search_k = 100 * n,
  ...
)
}
\arguments{
\item{cds}{A cell_data_set.}

\item{method}{The method for which the Annoy index was made. method
can be 'PCA', 'LSI','Aligned', 'tSNE', or 'UMAP'.}

\item{n}{An integer for the number of nearest neighbors to return for
each cell.}

\item{search_k}{An integer used to balance accuracy and speed. Larger
values give more accurate results.}

\item{...}{Parameters to pass through to annoy_search.}
}
\description{
Search the Annoy nearest neighbor index for cells near cells in cds.
}
