% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{set_assay_control}
\alias{set_assay_control}
\title{Check and set the assay_control list values.}
\usage{
set_assay_control(assay_control = list())
}
\arguments{
\item{assay_control}{Input control list.}
}
\value{
assay_control Output control list.
}
\description{
Check and set the assay_control list values.
}
