% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_neighbors.R
\name{set_cds_nn_index}
\alias{set_cds_nn_index}
\title{Set a nearest neighor index in the cell_data_set.}
\usage{
set_cds_nn_index(
  cds,
  reduction_method = c("UMAP", "PCA", "LSI", "Aligned", "tSNE"),
  nn_index,
  nn_control = list(),
  verbose = FALSE
)
}
\arguments{
\item{cds}{a cell_data_set in which to store the nearest
neighbor index.}

\item{reduction_method}{a string giving the reduced
dimension matrix used to make the nn_index nearest
neighbor index, and determines where the index is
stored in the cell_data_set.}

\item{nn_index}{a nearest neighbor index to store in cds.}

\item{nn_control}{a list of parameters used to make the nearest
neighbor index. This is used to identify where the index
is stored, based on the nearest neighbor method (annoy or hnsw)
and reduction_method. See the set_nn_control help for details.}

\item{verbose}{a boolean indicating whether to emit verbose output.}
}
\value{
a cell_data_set with the stored index.
}
\description{
Store the given nearest neighbor index
in the cell_data_set. The reduction_method parameter
tells set_cds_nn_index where in the cell_data_set to
store the index.
}
