% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{set_matrix_control}
\alias{set_matrix_control}
\title{Verify and set the matrix_control parameter list.}
\usage{
set_matrix_control(
  matrix_control = list(),
  matrix_control_default = list(),
  control_type = c("unrestricted", "counts", "mm", "pca")
)
}
\arguments{
\item{matrix_control}{Input control list.}

\item{matrix_control_default}{Input default control list.}

\item{control_type}{A string of either "unrestricted" or "pca". A control_type of "pca"
restricts certain list parameters.}
}
\value{
matrix_control Output control list.
}
\description{
Verifies and sets the list of parameter values
that is used to make the count matrix that is stored in the
cell_data_set or certain other matrices thare are used
during the Monocle3 run. To see the
default values,
call "set_matrix_control(matrix_control=list(matrix_class='BPCells', show_values=TRUE))".
"show_values=TRUE" can be used in functions that have the
matrix_control list parameter, in which case the function will
show the matrix_control values to be used and then stop.
}
\section{matrix_control parameters}{

\describe{
\item{matrix_class}{Specifies the matrix class to use for
matrix storage. The acceptable values are "dgCMatrix"
and "BPCells".}
\item{matrix_type}{Specifies whether to store the matrix
values as unsigned 32-bit integers
(matrix_type="uint32_t"), single precision "floats"
(matrix_type="float"), or double precision "doubles"
(matrix_type="double"). Unsigned 32-bit integers are
suitable for storing non-negative values such as counts.
"matrix_type" is used only for BPCells class matrices.}
\item{matrix_mode}{Specifies whether to store the BPCells
class matrix in memory (matrix_mode="mem") or on
disk (matrix_mode="dir"). "matrix_mode" is used only
for BPCells class matrices.}
\item{matrix_path}{Specifies the directory where the
BPCells on-disk matrix data are stored in a
sub-directory with a randomized name. The default is
in the directory where R is running. "matrix_path" is
used only for BPCells class matrices with
matrix_mode="dir". For example, if matrix_path is set
to "/tmp" Monocle3 will create a directory with a
name that has the form "monocle.bpcells.*.tmp" in
"/tmp". The asterisk represents a random string that
makes the names unique.}
\item{matrix_compress}{Specifies whether to use bit-packing
compression to store BPCells matrix values. This is
most effective for values stored as matrix_type="uint32_t"
but has some benefit for values stored as matrix_type="float"
and matrix_type="double". Compression reduces storage
requirements but increases run time. "matrix_compress" is
used only for BPCells class matrices.}
\item{matrix_buffer_size}{Specifies how many items of
data to buffer in memory before flushing to disk. This
is used for matrix_class="BPCells" with matrix_mode="dir".}
}
}

