% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-cell_data_set.R
\name{size_factors<-}
\alias{size_factors<-}
\title{Set the size factor values in the cell_data_set}
\usage{
size_factors(cds) <- value
}
\arguments{
\item{cds}{A cell_data_set object.}

\item{value}{the size factor values.}
}
\value{
An updated cell_data_set object
}
\description{
Set the size factor values in the cell_data_set
}
