% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-cell_data_set.R
\name{size_factors}
\alias{size_factors}
\title{Get the size factors from a cds object.}
\usage{
size_factors(cds)
}
\arguments{
\item{cds}{A cell_data_set object.}
}
\value{
An updated cell_data_set object
}
\description{
A wrapper around \code{colData(cds)$Size_Factor}
}
\examples{
  cds <- load_a549()
  size_factors(cds)
}
