% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_markers.R
\name{top_markers}
\alias{top_markers}
\title{Identify the genes most specifically expressed in specified groups of cells}
\usage{
top_markers(cds, group_cells_by = "cluster",
  genes_to_test_per_group = 50, reduction_method = "UMAP",
  expression_bins = 5, cores = 1)
}
\description{
Identify the genes most specifically expressed in specified groups of cells
}
