% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_markers.R
\name{top_markers}
\alias{top_markers}
\title{Identify the genes most specifically expressed in groups of cells}
\usage{
top_markers(
  cds,
  group_cells_by = "cluster",
  genes_to_test_per_group = 25,
  reduction_method = "UMAP",
  marker_sig_test = TRUE,
  reference_cells = NULL,
  speedglm.maxiter = 25,
  cores = 1,
  verbose = FALSE
)
}
\arguments{
\item{cds}{A cell_data_set object to calculate top markers for.}

\item{group_cells_by}{String indicating what to group cells by for
comparison. Default is "cluster".}

\item{genes_to_test_per_group}{Numeric, how many genes of the top ranked
specific genes by Jenson-Shannon to do the more expensive regression test
on.}

\item{reduction_method}{String indicating the method used for dimensionality
reduction. Currently only "UMAP" is supported.}

\item{marker_sig_test}{A flag indicating whether to assess the discriminative
power of each marker through logistic regression. Can be slow, consider
disabling to speed up top_markers().}

\item{reference_cells}{If provided, top_markers will perform the marker
significance test against a "reference set" of cells. Must be either a list
of cell ids from colnames(cds), or a positive integer. If the latter, top_markers()
will randomly select the specified number of reference cells. Accelerates
the marker significance test at some cost in sensitivity.}

\item{speedglm.maxiter}{Maximum number of iterations allowed for fitting GLM
models when testing markers for cell group.}

\item{cores}{Number of cores to use.}

\item{verbose}{Whether to print verbose progress output.}
}
\value{
a data.frame where the rows are genes and the columns ars
\itemize{
\item gene_id  vector of gene names.
\item gene_short_name vector of gene short names.
\item cell_group vector of cell groups to which the gene is assigned.
\item marker_score vector of marker scores. The value ranges from 0 to 1.
\item mean_expression vector of mean normalized expression of the gene in the cell group.
\item fraction_expressing vector of fraction of cells expressing gene within each cell group.
\item specificity vector of measure of how specific the gene's expression is to the cell group. The is based on the Jensen-Shannon divergence. The value ranges from 0 to 1.
\item pseudo_R2 vector of measure of how well the gene expression model fits the categorical data relative to the null model. The value ranges from 0 to 1.
\item marker_test_p_value vector of likelihood ratio p-value.
\item marker_test_q_value vector of likelihood ratio q-value.
}
}
\description{
Identify the genes most specifically expressed in groups of cells
}
