% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_markers.R
\name{top_markers}
\alias{top_markers}
\title{Identify the genes most specifically expressed in specified groups of cells}
\usage{
top_markers(cds, group_cells_by = "cluster",
  genes_to_test_per_group = 25, reduction_method = "UMAP",
  marker_sig_test = TRUE, reference_cells = NULL, cores = 1)
}
\arguments{
\item{cds}{A cell_data_set object to calculate top markers for.}

\item{group_cells_by}{String indicating what to group cells by for
comparison. Default is "cluster".}

\item{genes_to_test_per_group}{Numeric, how many genes of the top ranked
specific genes by Jenson-Shannon to do the more expensive regression test
on.}

\item{reduction_method}{String indicating the method used for dimensionality
reduction. Currently only "UMAP" is supported.}

\item{cores}{Number of cores to use.}

\item{expression_bins}{?}
}
\description{
Identify the genes most specifically expressed in specified groups of cells
}
